% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dependencies}
\alias{check_dependencies}
\title{Check SCpubr current function dependencies.}
\usage{
check_dependencies(
  function_name = NULL,
  return_dependencies = FALSE,
  bioconductor_repo = "https://bioconductor.org/packages/release/bioc",
  cran_repo = "http://cran.us.r-project.org"
)
}
\arguments{
\item{function_name}{\strong{\code{\link[base]{character}}} | Name of an exported function from SCpubr. If NULL, return all functions.}

\item{return_dependencies}{\strong{\code{\link[base]{logical}}} | Whether to have the dependencies as an output object instead of a printed message.}

\item{bioconductor_repo}{\strong{\code{\link[base]{bioconductor_repo}}} | Link to the repository that should be used to check \strong{\code{Bioconductor}} packages.}

\item{cran_repo}{\strong{\code{\link[base]{cran_repo}}} | Link to the repository that should be used to check \strong{\code{CRAN}} packages.}
}
\value{
None
}
\description{
Check SCpubr current function dependencies.
}
\examples{
\donttest{
# See all dependencies.
SCpubr::check_dependencies()

# See the dependencies for a single package.
SCpubr::check_dependencies(function_name = "do_DimPlot")
}
}
