% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_length}
\alias{check_length}
\title{Check length of parameters compared to features.}
\usage{
check_length(
  vector_of_parameters,
  vector_of_features,
  parameters_name,
  features_name
)
}
\arguments{
\item{vector_of_parameters}{Vector of parameters to test.}

\item{vector_of_features}{Vector of features to test against.}

\item{parameters_name}{Name of the parameters variable.}

\item{features_name}{Name of the features variable.}
}
\description{
Check length of parameters compared to features.
}
\examples{
\dontrun{
TBD
}
}
