% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_BarPlot.R
\name{do_BarPlot}
\alias{do_BarPlot}
\title{Create Bar Plots.}
\usage{
do_BarPlot(
  sample,
  group.by,
  order = TRUE,
  split.by = NULL,
  position = "stack",
  font.size = 14,
  font.type = "sans",
  legend.position = "bottom",
  rotate_x_axis_labels = FALSE,
  legend.title = NULL,
  xlab = NULL,
  ylab = NULL,
  colors.use = NULL,
  flip = FALSE,
  plot.title = NULL,
  plot.subtitle = NULL,
  plot.caption = NULL
)
}
\arguments{
\item{sample}{An object of class Seurat, generated by \link[Seurat]{CreateSeuratObject}.}

\item{group.by}{Character. Metadata variable to generate the counts of. Has to be either a character or a factor column.}

\item{order}{Logical. Whether to order the results in descending order of counts.}

\item{split.by}{Character. Further metadata variable to group the results by.}

\item{position}{Character. Either "stack" or "fill".}

\item{font.size}{\link[base]{numeric}. Overall font.size of the plot.}

\item{font.type}{\link[base]{character}. Base font for the plot. One of mono, serif or sans.}

\item{legend.position}{\link[base]{character}. Position of the legend in the plot. One of: top, bottom, left, right or none (if no legend is desired).}

\item{rotate_x_axis_labels}{\link[base]{logical}. Whether to rotate X axis labels.}

\item{legend.title}{TBD.}

\item{xlab, ylab}{\link[base]{character}. Titles for the X and Y axis.}

\item{colors.use}{TBD.}

\item{flip}{TBD.}

\item{plot.title, plot.subtitle, plot.caption}{\link[base]{character}. Title, subtitle or caption to use in the plot.}
}
\value{
A ggplot2 object containing a Bar plot.
}
\description{
Create Bar Plots.
}
\examples{
\dontrun{
  # Basic bar plot showing the number of items in a variable and a title.
  # Legend T/F toggles removes or keeps the legend.
  p <- SCpubr::do_BarPlot(sample = sample,
                          features = "seurat_clusters",
                          legend = F,
                          plot.title = "Number of cells per cluster")

  # Same but horizontal bars.
  p <- SCpubr::do_BarPlot(sample = sample,
                          features = "seurat_clusters",
                          legend = F,
                          horizontal = T)

  # Grouping by a second variable.
  # Position = stack will compute as separate bars.
  p <- SCpubr::do_BarPlot(sample,
                          features = "orig.ident",
                          group.by = "seurat_clusters",
                          position = "stack",
                          legend = T,
                          horizontal = F)
  # Position = fill will compute the proportions of each group in the feature.
  p <- SCpubr::do_BarPlot(sample,
                          features = "orig.ident",
                          group.by = "seurat_clusters",
                          position = "fill",
                          legend = T,
                          horizontal = F)

  # Grouping by a second variable and reordering based on an item in the second variable.
  # Reordering with position = fill will order the bars by descending proportion of order.by.
  p <- SCpubr::do_BarPlot(sample,
                          features = "modified_orig.ident",
                          group.by = "modified_seurat_clusters",
                          plot.title = "Number of cells per cluster and sample",
                          order.by = "1",
                          position = "fill",
                          legend = T,
                          horizontal = F)
  # Reordering with position = stack will order the bars by descending number of order.by.
  p <- SCpubr::do_BarPlot(sample,
                          features = "modified_orig.ident",
                          group.by = "modified_seurat_clusters",
                          plot.title = "Number of cells per sample",
                          order.by = "1",
                          position = "stack",
                          legend = T,
                          horizontal = F)

  # Adding custom colors.
  # Colors must be a named vector of HEX codes of the same lengths as bars/groups to color.
  colors <- c("0" = "#001219",
              "1" = "#005f73",
              "2" = "#0a9396",
              "3" = "#94d2bd",
              "4" = "#e9d8a6",
              "5" = "#ee9b00",
              "6" = "#ca6702",
              "7" = "#bb3e03",
              "8" = "#ae2012",
              "9" = "#9b2226")

  p <- SCpubr::do_BarPlot(sample = sample,
                          features = "seurat_clusters",
                          legend = F,
                          plot.title = "Number of cells per cluster",
                          horizontal = T,
                          colors.use = colors)
}
}
