% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ChordDiagramPlot.R
\name{do_ChordDiagramPlot}
\alias{do_ChordDiagramPlot}
\title{Generate a Chord diagram.}
\usage{
do_ChordDiagramPlot(
  sample = NULL,
  from = NULL,
  to = NULL,
  colors.from = NULL,
  colors.to = NULL,
  big.gap = 10,
  small.gap = 1,
  link.border.color = NA,
  link.border.width = 1,
  highlight_group = NULL,
  alpha.highlight = 25,
  link.sort = NULL,
  link.decreasing = TRUE,
  z_index = FALSE,
  self.link = 1,
  symmetric = FALSE,
  directional = 1,
  direction.type = c("diffHeight", "arrows"),
  link.arr.type = "big.arrow",
  scale = FALSE,
  alignment = "default",
  annotationTrack = c("grid", "axis"),
  padding_labels = 4,
  ...
)
}
\arguments{
\item{sample}{Seurat object.}

\item{from}{Character. Categorical metadata variable. First group to be used as origin of the interaction.}

\item{to}{Character. Categorical metadata variable. Last group to be used as the end of the interaction.}

\item{colors.from, colors.to}{Named vector of colors corresponding to the unique values of "from" and "to".}

\item{big.gap}{Numeric. Space between the groups in from and to.}

\item{small.gap}{Numeric. Space within the groups.}

\item{link.border.color}{Character. Color for the border of the links. NA = no color.}

\item{link.border.width}{Numeric. Width of the border line of the links.}

\item{highlight_group}{Character. A value from from that will be used to highlight only the links coming from it.}

\item{alpha.highlight}{Numeric. A value between 00 (double digits) and 99 to depict the alpha of the highlighted links. No transparency needs "FF"}

\item{link.sort}{pass to \code{\link[circlize]{chordDiagramFromMatrix}} or \code{\link[circlize]{chordDiagramFromDataFrame}}
}

\item{link.decreasing}{pass to \code{\link[circlize]{chordDiagramFromMatrix}} or \code{\link[circlize]{chordDiagramFromDataFrame}}
}

\item{z_index}{Logical. Whether to bring the bigger links to the top.}

\item{self.link}{Numeric. 1 to allow self linking and 2 to prevent them.}

\item{symmetric}{pass to \code{\link[circlize]{chordDiagramFromMatrix}}
}

\item{directional}{Numeric. 0 for non-directional data, 1 for from to to, -1 for to to from and 2 for bidirectional.}

\item{direction.type}{Character. One of diffHeight or arrows or both as a character vector.}

\item{link.arr.type}{Character. One of triangle or big.arrow.}

\item{scale}{Logical. Whether to put all nodes the same width.}

\item{alignment}{Character. One of default, vertical or horizontal. Default will allow circlize to put the angle as it wishes, vertical centers the symmetry on the Y axis and horizontal on the X axis.}

\item{annotationTrack}{pass to \code{\link[circlize]{chordDiagramFromMatrix}} or \code{\link[circlize]{chordDiagramFromDataFrame}}
}

\item{padding_labels}{Numeric. Number of extra padding of the labels so that they do not overlap with the scales.}

\item{...}{For internal use only.}
}
\value{
A circlize plot.
}
\description{
Generate a Chord diagram.
}
\examples{
\dontrun{
TBD
}
}
