% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ColorPalette.R
\name{do_ColorPalette}
\alias{do_ColorPalette}
\title{Generate color scales based on a value.}
\usage{
do_ColorPalette(
  colors.use,
  n = 12,
  opposite = FALSE,
  adjacent = FALSE,
  triadic = FALSE,
  split_complementary = FALSE,
  tetradic = FALSE,
  square = FALSE,
  complete_output = FALSE,
  plot = FALSE
)
}
\arguments{
\item{colors.use}{One color upon which generate the color scale. Can be a name or a HEX code.}

\item{n}{Numeric. Number of colors to include in the color wheel. Use it when all other options are FALSE, otherwise, it becomes 12.}

\item{opposite}{Logical. Return the opposing color to the one provided.}

\item{adjacent}{Logical. Return the adjacent colors to the one provided.}

\item{triadic}{Logical. Return the triadic combination of colors to the one provided.}

\item{split_complementary}{Logical. Return the split complementary combination of colors to the one provided.}

\item{tetradic}{Logical. Return the tetradic combination of colors to the one provided.}

\item{square}{Logical. Return the square combination of colors to the one provided.}

\item{complete_output}{Logical. Runs all the previous options and returns all the outputs as a list that contains all color vectors, all plots and a combined plot with everything.}

\item{plot}{Logical. Whether to also return a plot displaying the values instead of a vector with the color.}
}
\value{
A character vector with the desired color scale.

A ggplot2 object containing a Dot Plot.
}
\description{
This function is an adaptation of \link[colortools]{setColors}, \link[colortools]{complementary}, \link[colortools]{adjacent}, \link[colortools]{triadic}, \link[colortools]{split_complemmentary}, \link[colortools]{tetradic} and \link[colortools]{square}.
As the package was removed from CRAN on 23-06-2022, this utility function came to existence in order to cover the gap. It is, on its basis,
an adaptation of the package into a single function. Original code, developed by Gaston Sanchez, can be found in: \url{https://github.com/gastonstat/colortools}
}
\examples{
\dontrun{
TBD
}
}
