% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_CopyNumberVariantPlot.R
\name{do_CopyNumberVariantPlot}
\alias{do_CopyNumberVariantPlot}
\title{Display CNV scores from inferCNV as Feature Plots.}
\usage{
do_CopyNumberVariantPlot(
  sample,
  infercnv_object,
  group.by = NULL,
  using_metacells = FALSE,
  metacell_mapping = NULL,
  chromosome_focus = NULL,
  legend.type = "colorbar",
  legend.position = "bottom",
  legend.length = 30,
  legend.width = 1,
  legend.framewidth = 1.5,
  legend.tickwidth = 1.5,
  legend.framecolor = "grey50",
  legend.tickcolor = "white",
  font.size = 14,
  pt.size = 1,
  font.type = "sans",
  border.size = 1.5,
  rotate_x_axis_labels = TRUE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{infercnv_object}{Output inferCNV object run on the same Seurat object.}

\item{group.by}{Metadata variable to group cells by. If not provided, defaults to active identitites.}

\item{using_metacells}{Logical. Whether inferCNV was run using metacells or not.}

\item{metacell_mapping}{Vector or cell - metacell mapping.}

\item{chromosome_focus}{Character. Region stating which chromosome to plot. Eg: 1p, 19q. NULL will plot all regions.}

\item{legend.type}{Character. Type of legend to display. One of: normal, colorbar, colorsteps.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.length, legend.width}{Length and width of the legend. Will adjust automatically depending on legend side.}

\item{legend.framewidth, legend.tickwidth}{Width of the lines of the box in the legend.}

\item{legend.framecolor, legend.tickcolor}{Color of the lines of the box in the legend.}

\item{pt.size}{Size of the dots in the dotplot pltos.}

\item{font.type}{Character. Base font for the plot. One of mono, serif or sans.}

\item{border.size}{Thickness of the border around cells in all plots.}

\item{rotate_x_axis_labels}{Logical. Rotate x axis labels 90 degrees in the dotplot plots.}
}
\value{
A list containing Feature Plots for different chromosome regions and corresponding dot plots by groups..
}
\description{
Display CNV scores from inferCNV as Feature Plots.
}
\examples{
\dontrun{
  # This function expects that you have run inferCNV on your own and you have access to the output object.

  # Using a normal run - compute for all chromosome arms.
  out <- SCpubr::do_CopyNumberVariantPlot(sample = sample,
                                          infercnv_object = infercnv_object,
                                          using_metacells = FALSE)
  # Retrieve the UMAP for 1p region.
  out$`1p_umap`
  # Retrieve the dot plot for 1p region.
  out$`1p_dotplot`

  # Compute for a single chromosoome.
  out <- SCpubr::do_CopyNumberVariantPlot(sample = sample,
                                          infercnv_object = infercnv_object,
                                          using_metacells = FALSE,
                                          chromosome_focus = "2")

  # Compute for a run using metacells.
  ## metacell_mapping has to be a vector that, for each cell, tells which is the corresponding metacell.
  out <- SCpubr::do_CopyNumberVariantPlot(sample = sample,
                                          infercnv_object = infercnv_object_metacells,
                                          using_metacells = TRUE,
                                          metacell_mapping = metacell_mapping)

}
}
