% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_EnrichedTermPlot.R
\name{do_EnrichedTermPlot}
\alias{do_EnrichedTermPlot}
\title{Compute the enriched terms for a given list of genes.}
\usage{
do_EnrichedTermPlot(
  genes,
  dbs_use = NULL,
  ncol = NULL,
  nchar_wrap = 60,
  nterms = 5,
  size = 4,
  fontsize = 14,
  site = "Enrichr"
)
}
\arguments{
\item{genes}{Character vector with the genes to query.}

\item{dbs_use}{Character vector with the desired EnrichR databases to query. This can be consulted with \code{enrichR::listEnrichrDbs()}. Alternatively, one of the following values:
\itemize{
\item "A": Will perform a query to 4 databases for cell types (Azimuth, Descartes, PanglaoDB and Descartes) and 4 databases for functional terms (MsigDB, GO-BP, GO-MF and KEGG). This is the default option if this parameter is not provided.
\item "B": Performs a query for the cell type databases (Azimuth, Descartes, PanglaoDB and Descartes).
\item "C": Performs a query for the functional terms (MsigDB, GO-BP, GO-MF and KEGG).
}}

\item{ncol}{Number of columns to group the output plots. Defaults to a around half of the length of the databases provided, if this is more than 1.}

\item{nchar_wrap}{Number of characters to use as a limit to wrap the term names. The higher this value, the longer the lines would be for each term in the plots. Defaults to 60.}

\item{nterms}{Number of terms to report for each database. Terms are arranged by adjusted p-value and selected from lowest to highest. Defaults to 5.}

\item{size}{Size of the dots. Defaults to 4.}

\item{fontsize}{Base font size for the plot. Defaults to 14.}

\item{site}{Site to query the genes against. Can be one of: "Enrichr", "FlyEnrichr", "WormEnrichr", "YeastEnrichr", "FishEnrichr".}
}
\value{
A ggplot2 object with enriched terms.
}
\description{
Compute the enriched terms for a given list of genes.
}
