% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_EnrichmentHeatmap.R
\name{do_EnrichmentHeatmap}
\alias{do_EnrichmentHeatmap}
\title{Create enrichment scores heatmaps.}
\usage{
do_EnrichmentHeatmap(
  sample,
  list_genes,
  group.by = NULL,
  split.by = NULL,
  split.horizontal = TRUE,
  column_title = NULL,
  row_title = NULL,
  verbose = FALSE,
  transpose = FALSE,
  cluster_cols = TRUE,
  cluster_rows = FALSE,
  legend_name = "Enrichment",
  row_names_rot = 0,
  column_names_rot = 90,
  colors.use = NULL
)
}
\arguments{
\item{sample}{Your Seurat object.}

\item{list_genes}{Named list of genes to query to the Seurat object.}

\item{group.by}{Metadata variable to group the values by.}

\item{split.by}{Metadata variable to split the values of group.by into as many individual heatmaps as unique values in split.by.}

\item{split.horizontal}{Logical. Join the individual heatmaps generated by split.by horizontally (TRUE) or vertically (FALSE).}

\item{column_title, row_title}{Title for the columns and rows. Only works with single heatmaps.}

\item{verbose}{Extended output.}

\item{transpose}{Logical. Transpose the resulting heatmap.}

\item{cluster_cols, cluster_rows}{Logical. Cluster the columns or rows.}

\item{legend_name}{Text for the legend title.}

\item{column_names_rot, row_names_rot}{Numeric. Degree in which to}

\item{colors.use}{Vector of 2 colors to use to generate the color scale.}
}
\value{
A ComplexHeatmap object.
}
\description{
This function computes the enrichment scores for the cells using \link[Seurat]{AddModuleScore} and then aggregates the scores by the metadata variables provided by the user and displays it as a heatmap, computed by \link[ComplexHeatmap]{Heatmap}.
}
