% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_GeyserPlot.R
\name{do_GeyserPlot}
\alias{do_GeyserPlot}
\title{Generate a Geyser plot.}
\usage{
do_GeyserPlot(
  sample,
  features,
  assay = NULL,
  slot = "data",
  group.by = NULL,
  split.by = NULL,
  symmetrical_scale = FALSE,
  scale_type = "continuous",
  color.by = NULL,
  order_by_mean = TRUE,
  plot_cell_borders = TRUE,
  jitter = 0.45,
  pt.size = 1,
  border.size = 1.5,
  border.color = "black",
  legend.position = "bottom",
  legend.width = 1,
  legend.length = 20,
  legend.framewidth = 1.5,
  legend.tickwidth = 1.5,
  legend.framecolor = "grey50",
  legend.tickcolor = "white",
  legend.type = "colorbar",
  font.size = 14,
  font.type = "sans",
  rotate_x_axis_labels = FALSE,
  viridis_color_map = "D",
  colors.use = NULL,
  na.value = "grey75",
  legend.ncol = NULL,
  legend.nrow = NULL,
  legend.icon.size = 4,
  legend.byrow = FALSE,
  legend.title = NULL,
  show_legend = TRUE,
  plot.title = NULL,
  plot.subtitle = NULL,
  plot.caption = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Character. Features to plot.}

\item{assay}{Character. Assay to use. Defaults to active assay if not.}

\item{slot}{Character. Slot to use. Defaults to data slot if not.}

\item{group.by}{Character. Metadata variable to group the data by.}

\item{split.by}{Character. Metadata variable to further split the plot by.}

\item{symmetrical_scale}{Logical. Whether you want a continuous scale that is symmetrical on both ends and the intensity of the colors in each end matches (blue to red).}

\item{scale_type}{Character. Either continuous or categorical.}

\item{color.by}{Character. Vector of colors to color the groups by if scale_type is categorical.}

\item{order_by_mean}{Logical. Whether to order the groups by the mean of the data (highest to lowest).}

\item{plot_cell_borders}{Logical. Whether to plot border around the cells.}

\item{jitter}{Numeric. Amount of jitter in the plot along the X axis. The lower the value, the more compacted the dots are.}

\item{pt.size}{Numeric. Size of the cells.}

\item{border.size}{Numeric. Thickness of the border around the cells.}

\item{border.color}{Character. Color for the border of the cells.}

\item{legend.position}{Character. Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.length, legend.width}{Numeric. Length and width of the legend. Will adjust automatically depending on legend side.}

\item{legend.framewidth, legend.tickwidth}{Numeric. Width of the lines of the box in the legend.}

\item{legend.framecolor, legend.tickcolor}{Character. Color of the lines of the box in the legend.}

\item{legend.type}{Character. Type of legend to display. One of: normal, colorbar, colorsteps.}

\item{font.size}{Numeric. Overall font size of the plot.}

\item{font.type}{Character. Font family for the plot: sans, mono, serif.}

\item{rotate_x_axis_labels}{Logical. Whether to rotate X axis labels.}

\item{viridis_color_map}{Character. Viridis color map to use. One of: A, B, C, D, E, F, G, H.}

\item{colors.use}{Character. Named vector of colors to use. Has to match the unique values of group.by or color.by (if used) when scale_type is set to categorical.}

\item{na.value}{Character. Color for NAs.}

\item{legend.ncol, legend.nrow}{Numeric. Number of rows/columns of a categorical legend.}

\item{legend.icon.size}{Numeric. Size of the dots in a categorical legend.}

\item{legend.byrow}{Logical. Whether to fill the legend by rows instead of columns in a categorical legend.}

\item{legend.title}{Character. Title for the legend.}

\item{show_legend}{Logical. Whether to display the legend. Might be redundant depending on the setups.}

\item{plot.title, plot.subtitle, plot.caption}{Character. Title, Subtitle and caption to use in the plot.}

\item{xlab, ylab}{Character. Titles for the X and Y axis.}
}
\value{
Either a plot of a list of plots, depending on the number of features provided.
}
\description{
A Geyser plot is a custom plot in which we plot continuous values on the Y axis grouped by a categorical value in the X. This is plotted as a dot plot, jittered so that the dots span
all the way to the other groups. On top of this, the mean and .66 and .95 of the data is plotted, depicting the overall distribution of the dots. The cells can, then, be colored by
a continuous variable (same as Y axis or different) or a categorical one (same as X axis or different).
}
\examples{
\dontrun{
TBD
}
}
