% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_GroupedGOPlot.R
\name{do_GroupedGOPlot}
\alias{do_GroupedGOPlot}
\title{Compute an overview of the GO terms associated with the input list of genes.}
\usage{
do_GroupedGOPlot(
  genes,
  org.db,
  levels.use = NULL,
  GO_ontology = "BP",
  min.overlap = 3,
  flip = TRUE,
  legend.position = "none",
  heatmap_gap = 1,
  cluster_rows = TRUE,
  cluster_columns = TRUE,
  cell_size = 5,
  reverse.levels = TRUE,
  colors.use = c("white", "#29353d"),
  rotate_x_axis_labels = 45,
  font.size = 10
)
}
\arguments{
\item{genes}{\strong{\code{\link[base]{character}}} | Vector of gene symbols to query for functional annotation.}

\item{org.db}{\strong{\code{OrgDB}} | Database object to use for the query.}

\item{levels.use}{\strong{\code{\link[base]{character}}} | Vector of characters corresponding to the GO ontology levels to plot. If NULL will compute all recursively until there are no results.}

\item{GO_ontology}{\strong{\code{\link[base]{character}}} | GO ontology to use. One of:
\itemize{
\item \emph{\code{BP}}: For \strong{B}iological \strong{P}rocess.
\item \emph{\code{MF}}: For \strong{M}olecular \strong{F}unction.
\item \emph{\code{CC}}: For \strong{C}ellular \strong{C}omponent.
}}

\item{min.overlap}{\strong{\code{\link[base]{numeric}}} | Filter the output result to the terms which are supported by this many genes.}

\item{flip}{\strong{\code{\link[base]{logical}}} | Whether to invert the axis of the displayed plot.}

\item{legend.position}{\strong{\code{\link[base]{character}}} | Position of the legend in the plot. One of:
\itemize{
\item \emph{\code{top}}: Top of the figure.
\item \emph{\code{bottom}}: Bottom of the figure.
\item \emph{\code{left}}: Left of the figure.
\item \emph{\code{right}}: Right of the figure.
\item \emph{\code{none}}: No legend is displayed.
}}

\item{heatmap_gap}{\strong{\code{\link[base]{numeric}}} | Gap in cm between heatmaps.}

\item{cell_size}{\strong{\code{\link[base]{numeric}}} | Size of each cell in the heatmap.}

\item{reverse.levels}{\strong{\code{\link[base]{logical}}} | Whether to place the higher levels first when computing the joint heatmap.}

\item{colors.use}{\strong{\code{\link[base]{character}}} | Vector of 2 colors to use in the heatmap. The first will correspond to the empty values and the second one to the genes present in the terms.}

\item{rotate_x_axis_labels}{\strong{\code{\link[base]{numeric}}} | Degree to rotate the X labels. One of: 0, 45, 90.}

\item{font.size}{\strong{\code{\link[base]{numeric}}} | Overall font size of the plot. All plot elements will have a size relationship with this font size.}
}
\value{
A list containing all the matrices for the respective GO levels and all the individual and combined heatmaps.
}
\description{
Compute an overview of the GO terms associated with the input list of genes.
}
