% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_GroupwiseDEPlot.R
\name{do_GroupwiseDEPlot}
\alias{do_GroupwiseDEPlot}
\title{Compute a heatmap with the results of a group-wise DE analysis.}
\usage{
do_GroupwiseDEPlot(
  sample,
  de_genes,
  group.by = "seurat_clusters",
  viridis_map_pvalues = "B",
  viridis_map_logfc = "D",
  viridis_map_expression = "G",
  heatmap.legend.length = 75,
  heatmap.legend.width = 5,
  heatmap.legend.framecolor = "black",
  top_genes = 5,
  scale_direction = -1,
  row_title_p_values = "",
  row_title_logfc = "Clusters",
  row_title_expression = c(""),
  column_title = "DE genes",
  heatmap_gap = 0.5,
  legend_gap = 1,
  assay = "SCT",
  slot = "data",
  legend.position = "bottom",
  row_names_side = "right",
  row_title_side = "left",
  row_title_rotation = 90,
  cell_size = 5
)
}
\arguments{
\item{sample}{Seurat objects.}

\item{de_genes}{DE genes matrix resulting of running \code{Seurat::FindAllMarkers()}.}

\item{group.by}{Character. Metadata variable (s) that was used with \code{Seurat::FindAllMarkers()}. Generally, it is "seurat_clusters", but it is normally the one used as current identities of the Seurat object. One can add further variables in a vector, and they will be plotted on top of each other.}

\item{viridis_map_pvalues, viridis_map_logfc, viridis_map_expression}{Character. Viridis color map for the heatmap of p-values, logFC or expression. One of A, B, C, D, E, F, G, H.}

\item{heatmap.legend.length, heatmap.legend.width}{Numeric. Width and length of the legend in the heatmap.}

\item{heatmap.legend.framecolor}{Character. Color of the edges and ticks of the legend in the heatmap.}

\item{top_genes}{Numeric. Top N DE genes by group to retrieve.}

\item{scale_direction}{Numeric. Direction of the viridis scales. Either -1 or 1.}

\item{row_title_p_values}{Character. Row title for the p-value heatmap. Blank by default.}

\item{row_title_logfc}{Character. Row title for the logfc heatmap. Clusters by default.}

\item{row_title_expression}{Character. Vector of titles of equal length as group.by.}

\item{column_title}{Character. Title for the column.}

\item{heatmap_gap, legend_gap}{Numeric. Gap in cm between legends or heatmaps.}

\item{assay}{Character. Assay used to compute the DE genes.}

\item{slot}{Character. Slot from the assay used to compute the DE genes. Normally, this is "data".}

\item{legend.position}{Character. Where to place the legends. Either "top", "botom", "right", "left".}

\item{row_names_side}{Character. Either left or right.}

\item{row_title_side}{Character. Either left or right.}

\item{row_title_rotation}{Numeric. Degree of rotation of the row titles.}

\item{cell_size}{Numeric. Size of the cells in the heatmap.}
}
\value{
A heatmap composed of 3 main panels: -log10(adjusted p-value), log2(FC) and mean expression by cluster.
}
\description{
Compute a heatmap with the results of a group-wise DE analysis.
}
\examples{
\dontrun{
  # Seurat sample.
  # sample <- your_seurat_object

  # Set the identities correctly.
  # Seurat::Idents(sample) <- sample$seurat_clusters

  # Compute DE genes and transform to a tibble.
  # de_genes <- tibble::tibble(Seurat::FindAllMarkers(object = sample))

  # Default output.
  p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
                                  de_genes = de_genes)

  p

  # Increase the number of top DE genes by cluster.
  p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
                                  de_genes = de_genes,
                                  top_genes = 10)

  p

  # Modify the row and column titles and the rotation.
  p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
                                  de_genes = de_genes,
                                  column_title = "Title A",
                                  row_title_p_values = "Title B",
                                  row_title_logfc = "Title C",
                                  row_title_expression = "Title D",
                                  row_title_rotation = 0)

  p

  sample$modified_orig.ident <- sample(x = c("Sample_A", "Sample_B", "Sample_C"),
                                       size = ncol(sample),
                                       replace = T,
                                       prob = c(0.2, 0.7, 0.1))

  # Add more layers of mean expression with group.by.
  p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
                                  de_genes = de_genes,
                                  group.by = c("seurat_clusters",
                                               "modified_orig.ident",
                                               "orig.ident"),
                                  row_title_expression = c("",
                                                           "Title A",
                                                           "Title B"))

  p

  # Change the viridis scales.
  p <- SCpubr::do_GroupwiseDEPlot(sample = sample,
                                  de_genes = de_genes,
                                  viridis_map_pvalues = "C",
                                  viridis_map_logfc = "E",
                                  viridis_map_expression = "D")

  p
}
}
