% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_PackageReport}
\alias{do_PackageReport}
\title{Generate a status report of SCpubr and its dependencies.}
\usage{
do_PackageReport(startup = FALSE, extended = FALSE)
}
\arguments{
\item{startup}{\strong{\code{\link[base]{logical}}} | Whether the message should be displayed at startup, therefore, also containing welcoming messages and tips. If \strong{\code{FALSE}}, only the report itself will be printed.}

\item{extended}{\strong{\code{\link[base]{logical}}} | Whether the message should also include installed packages, current and available version, and which \strong{\code{SCpubr}} functions can be used with the currently installed packages.}
}
\value{
None
}
\description{
This function generates a summary report of the installation status of SCpubr, which packages are still missing and which functions can or can not currently be used.
}
\examples{

\donttest{
# Print a package report.
SCpubr::do_PackageReport(startup = FALSE, extended = FALSE)
}
}
