% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_PTEA.R
\name{do_PTEA}
\alias{do_PTEA}
\title{Retrieve statistically enriched cells for a list of genes in gradient-like enrichment cases.}
\usage{
do_PTEA(
  sample,
  markers,
  list.name,
  group.by = NULL,
  colors.use = NULL,
  compute_enrichment = FALSE,
  FDR_cutoff = 0.05,
  number_comparisons = 1,
  verbose = T
)
}
\arguments{
\item{sample}{Seurat object.}

\item{markers}{Named list of marker genes. Can contain multiple lists, the important point here is that each list has to be named.}

\item{list.name}{Name of the list of markers to use.}

\item{group.by}{Variable you want the cells to be colored for in the output DimPlot.}

\item{colors.use}{Vector of named HEX values to color the cells. It has to match the number of unique values in either \code{Seurat::Idents(sample)} or the group.by variable.}

\item{compute_enrichment}{Whether to compute the enrichment scores for the requested list of genes using \link[Seurat]{AddModuleScore}.}

\item{FDR_cutoff}{FDR cutoff to apply (ranging from 0 to 1).}

\item{number_comparisons}{Number of different lists that are going to be queried to the same cells. FDR value will be divided by this.}

\item{verbose}{Defaults to TRUE. It will provide different print statements. Progress bars can not be suppressed by this.}
}
\value{
A list containing the plots and the surpassed cells, together with the p-value matrix.
}
\description{
Designed for 10X datasets. The following analysis aims to find a way to statistically select the cells that are truly enriched in a list of marker genes.
For this, a permutation analysis approach is followed, in which we compute a null distribution by shuffling the expression values
of the genes queried for enrichment. Therefore, this disrupts the equation for the enrichment scores, since they are, in short,
the result of the difference in means between the expression values of the genes queried (which we disrupted) and the control
selected randomly by the function. This is done until we have one million values, which would act as the null distribution.
Once we have a null distribution and an empirical one, we can assess how extreme our values in the empirical distribution are with
respect to the values in the null distribution, being the p-value the proportion of values in the null distribution that are higher
than the given value from the empirical distribution you are querying. These p-values are, then, adjusted for multiple testing, selecting
a FDR cutoff of 0.05/n, being n the total number of lists of marker genes we are going to query to the same cells and use altogether
to define a labeling (for instance, if we query two lists for the same tumor bulk, it would be 0.05/2). This is doing to avoid
over inflation of the alpha error.
}
\details{
Estimated running time: 15 minutes.
}
\examples{
\dontrun{
TBD
}
}
