% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_SankeyPlot.R
\name{do_SankeyPlot}
\alias{do_SankeyPlot}
\title{Do Sankey or Alluvial plots.}
\usage{
do_SankeyPlot(
  sample,
  first_group,
  last_group,
  type = "sankey",
  middle_groups = NULL,
  width = 0.1,
  space = 0.05 * ncol(sample),
  position = "identity",
  node.fill = "white",
  node.color = "white",
  flow.alpha = 0.75,
  flow.color = "black",
  text_size = 3,
  text_color = "black",
  font.size = 14,
  font.type = "sans",
  smooth = 8,
  colors.first = NULL,
  colors.middle = NULL,
  colors.last = NULL,
  plot.title = NULL,
  plot.subtitle = NULL,
  plot.caption = NULL
)
}
\arguments{
\item{sample}{Seurat object.}

\item{first_group}{Character. Categorical metadata variable. First group of nodes of the sankey plot.}

\item{last_group}{Character. Categorical metadata variable. Last group of nodes of the sankey plot.}

\item{type}{Character. Type of plot. Either sankey or alluvial.}

\item{middle_groups}{Character. Categorical metadata variable. Vector of groups of nodes of the sankey plot.}

\item{width}{Numeric. Width of the nodes.}

\item{space}{Numeric. Vertical space between the nodes. It appears to be equal to a single cell. Use big numbers to see a difference (like, 1000 or 10000).}

\item{position}{Character. GGplot2 position.}

\item{node.fill}{Character. Color to fill the nodes.}

\item{node.color}{Character. Color for the contour of the nodes.}

\item{flow.alpha}{Character. Alpha of the connections.}

\item{flow.color}{Character. Color for the contour of the connections.}

\item{text_size}{Numeric. Size of the labels.}

\item{text_color}{Character. Color of the labels.}

\item{font.size}{Numeric. Overall font size of the plot.}

\item{font.type}{Character. Font family: either sans, mono or serif.}

\item{smooth}{Numeric. How smooth the connections are.}

\item{colors.first, colors.middle, colors.last}{Character. Named vector of colors equal to ALL unique values in first_group, middle_groups, or last_group.}

\item{plot.title, plot.subtitle, plot.caption}{Character. Title, Subtitle and caption to use in the plot.}
}
\value{
A ggplot2 object.
}
\description{
Do Sankey or Alluvial plots.
}
\examples{
\dontrun{
TBD
}
}
