% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_TFActivityPlot.R
\name{do_TFActivityPlot}
\alias{do_TFActivityPlot}
\title{Plot TF Activities from DecoupleR using Dorothea prior knowledge.}
\usage{
do_TFActivityPlot(
  sample,
  activities,
  n_tfs = 25,
  group.by = NULL,
  split.by = NULL,
  plot_FeaturePlots = FALSE,
  plot_Heatmaps = TRUE,
  plot_GeyserPlots = FALSE,
  geyser_color.by = NULL,
  row_title = NULL,
  column_title = NULL,
  transpose = FALSE,
  cluster_cols = TRUE,
  cluster_rows = TRUE,
  row_names_rot = 0,
  column_names_rot = 90,
  cell_size = 5,
  pt.size = 1,
  plot_cell_borders = TRUE,
  border.size = 1.5,
  na.value = "grey75",
  legend.position = "bottom",
  heatmap.legend.length = 75,
  heatmap.legend.width = 5,
  heatmap.legend.framecolor = "black",
  legend.width = 1,
  legend.length = 20,
  legend.framewidth = 1.5,
  legend.tickwidth = 1.5,
  legend.framecolor = "grey50",
  legend.tickcolor = "white",
  legend.type = "colorbar",
  font.size = 14,
  font.type = "sans",
  rotate_x_axis_labels = FALSE,
  symmetrical_scale = TRUE,
  geyser_order_by_mean = TRUE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{activities}{Result of running decoupleR method with dorothea regulon prior knowledge.}

\item{n_tfs}{Numeric. Number of top regulons to consider for downstream analysis.}

\item{group.by}{Character. Variable to group the output by. Defaults to current identities.}

\item{split.by}{Character. Optional variable to split the heatmaps by as well.}

\item{plot_FeaturePlots}{Logical. Compute output feature plots for each of the top regulons.}

\item{plot_Heatmaps}{Logical. Compute output heatmap showcasing the average TF activity per regulon and group.by variable.}

\item{plot_GeyserPlots}{Logical. Compute output dotplot for each of the top regulons and group.by variable.}

\item{geyser_color.by}{Character. Additional variable to color the Geyser plots by, as the Y axis and the color scale are repeated. Has to be a continuous variable.}

\item{row_title, column_title}{Character. Title for the rows in the heatmap.}

\item{transpose}{Logical. Whether to transpose the heatmap matrix.}

\item{cluster_cols, cluster_rows}{Logical. Whether to cluster rows and columns in the heatmap.}

\item{row_names_rot, column_names_rot}{Numeric. Angle of rotation of the column and row names. Suggested: either 0 or 90.}

\item{cell_size}{Numeric. Size of each cell in the heatmap.}

\item{pt.size}{Numeric. Size of the dots in the non-heatmap plots.}

\item{plot_cell_borders}{Logical. Whether to plot borders around cells in non-heatmap plots.}

\item{border.size}{Numeric. Size of the border.}

\item{na.value}{Character. Color for NA values.}

\item{legend.position}{Character. Position of the legend in the plots. One of: top, bottom, left, right.}

\item{heatmap.legend.length, heatmap.legend.width}{Numeric. Width and length of the legend in the heatmap.}

\item{heatmap.legend.framecolor}{Character. Color of the edges and ticks of the legend in the heatmap.}

\item{legend.length, legend.width}{Length and width of the legend for both feature plots and dot plots. Will adjust automatically depending on legend side.}

\item{legend.framewidth, legend.tickwidth}{Width of the lines of the box in the legend for both feature plots and dot plots.}

\item{legend.framecolor, legend.tickcolor}{Color of the lines of the box in the legend for both feature plots and dot plots.}

\item{legend.type}{Character. Type of legend to display. One of: normal, colorbar, colorsteps.}

\item{font.size}{Numeric. Base font.size of the figure.}

\item{font.type}{Character. Base font for the plot. One of mono, serif or sans.}

\item{rotate_x_axis_labels}{Logical. Whether to rotate the X axis names in the dot plot.}

\item{symmetrical_scale}{Logical. Whether the geyser and feature plot has a symmetrical color scale.}

\item{geyser_order_by_mean}{Logical. Whether to order the X axis by the mean of the values.}
}
\value{
A list containing several output plots according to the user's specifications.
}
\description{
Plot TF Activities from DecoupleR using Dorothea prior knowledge.
}
\examples{
\dontrun{
TBD
}
}
