% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_VlnPlot.R
\name{do_VlnPlot}
\alias{do_VlnPlot}
\title{Wrapper for \link[Seurat]{VlnPlot}.}
\usage{
do_VlnPlot(
  sample,
  features,
  group.by = NULL,
  split.by = NULL,
  legend = TRUE,
  cols = NULL,
  cols.group = NULL,
  cols.split = NULL,
  pt.size = 0,
  y_cut = NULL,
  plot_boxplot = TRUE,
  legend.position = "bottom",
  plot.title = "",
  xlab = "",
  ylab = "",
  axis.text.fonsize = 28,
  axis.title.fonsize = 28,
  plot.title.fonsize = 30,
  legend.text.fontsize = 20,
  legend.title.fontsize = 24,
  remove_x_axis = TRUE,
  remove_y_axis = FALSE,
  legend.ncol = 3
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Features to represent.}

\item{group.by}{Variable you want the cells to be colored for.}

\item{split.by}{Split into as many plots as unique values in the variable provided.}

\item{legend}{Whether to plot the legend or not.}

\item{cols}{From \link[Seurat]{DotPlot}: Colors to plot: the name of a palette from \code{RColorBrewer::brewer.pal.info}, a pair of colors defining a gradient, or 3+ colors defining multiple gradients (if split.by is set).}

\item{cols.group}{Vector of colors matching the unique values in group.by.}

\item{cols.split}{Vector of colors matching the unique values in split.by.}

\item{pt.size}{Size of points in the VlnPlot.}

\item{y_cut}{Vector with the values in which the Violins should be cut. Only works for one feature.}

\item{plot_boxplot}{Logical. Whether to plot a boxplot inside the violin or not.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{plot.title}{Title to use in the plot.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{axis.text.fonsize}{Modify the fontsize for axis texts.}

\item{axis.title.fonsize}{Modify the fontsize for axis titles.}

\item{plot.title.fonsize}{Modify the fontsize for the plot title.}

\item{legend.text.fontsize}{Modify the fontsize for the legend text.}

\item{legend.title.fontsize}{Modify the fontsize for the legend title.}

\item{remove_x_axis}{Remove X axis labels and ticks from the plot.}

\item{remove_y_axis}{Remove Y axis labels and ticks from the plot.}

\item{legend.ncol}{Number of columns in the legend.}
}
\value{
A ggplot2 object containing a Violin Plot.
}
\description{
Wrapper for \link[Seurat]{VlnPlot}.
}
\examples{
\dontrun{
TBD
}
}
