% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_VolcanoPlot.R
\name{do_VolcanoPlot}
\alias{do_VolcanoPlot}
\title{Compute a Volcano plot out of DE genes.}
\usage{
do_VolcanoPlot(
  sample,
  de_genes,
  pval_cutoff = 0.05,
  FC_cutoff = 2,
  pt.size = 2,
  border.size = 1.5,
  border.color = "black",
  font.size = 14,
  font.type = "sans",
  plot.title = NULL,
  plot.subtitle = NULL,
  plot.caption = NULL,
  plot_lines = TRUE,
  line_color = "grey75",
  line_size = 0.5,
  add_gene_tags = TRUE,
  order_tags_by = "both",
  n_genes = 5,
  use_labels = FALSE,
  colors.use = "steelblue"
)
}
\arguments{
\item{sample}{Seurat object.}

\item{de_genes}{Output of \code{Seurat::FindMarkers()}.}

\item{pval_cutoff}{Numeric. Cutoff for the p-value.}

\item{FC_cutoff}{Numeric. Cutoff for the avg_log2FC.}

\item{pt.size}{Numeric. Size for the dots.}

\item{border.size}{Numeric. Size for the border of the dots.}

\item{border.color}{Character. Color for the border of the dots.}

\item{font.size}{Numeric. Overall size of the font in the plot.}

\item{font.type}{Character. Type of font.}

\item{plot.title, plot.subtitle, plot.caption}{Character. Title, subtitle and caption for the plot.}

\item{plot_lines}{Logical. Whether to plot the division lines.}

\item{line_color}{Character. Color for the lines.}

\item{line_size}{Numeric. Size of the lines in the plot.}

\item{add_gene_tags}{Logical. Whether to plot the top genes.}

\item{order_tags_by}{Character. Either "both", "p_value" or "logfc".}

\item{n_genes}{Numeric. Number of top genes in each side to plot.}

\item{use_labels}{Logical. Whether to use labels instead of text for the tags.}

\item{colors.use}{Character. Color to generate a tetradic color scale with.}
}
\value{
A volcano plot as a ggplot2 object.
}
\description{
Compute a Volcano plot out of DE genes.
}
\examples{
\dontrun{
  # Generate a volcano plot.
  p <- SCpubr::do_VolcanoPlot(sample = sample,
                              de_genes = de_genes)
  p

  # Modify cutoffs.
  p <- SCpubr::do_VolcanoPlot(sample = sample,
                              de_genes = de_genes,
                              pval_cutoff = 1e-15,
                              FC_cutoff = 0.2)
  p

  # Modify number of gene tags.
  p <- SCpubr::do_VolcanoPlot(sample = sample,
                              de_genes = de_genes,
                              n_genes = 15)
  p
}
}
