% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doc_function}
\alias{doc_function}
\title{Mock function used to document all main function.}
\usage{
doc_function(
  sample,
  font.size,
  font.type,
  legend.type,
  legend.position,
  legend.framewidth,
  legend.tickwidth,
  legend.framecolor,
  legend.tickcolor,
  legend.length,
  legend.width,
  plot.title,
  plot.subtitle,
  plot.caption,
  assay,
  slot,
  reduction,
  viridis_color_map,
  raster,
  raster.dpi,
  plot_cell_borders,
  border.size,
  border.color,
  na.value,
  rotate_x_axis_labels,
  xlab,
  ylab
)
}
\arguments{
\item{sample}{An object of class Seurat, generated by \link[Seurat]{CreateSeuratObject}.}

\item{font.size}{\link[base]{numeric}. Overall font.size of the plot.}

\item{font.type}{\link[base]{character}. Base font for the plot. One of mono, serif or sans.}

\item{legend.type}{\link[base]{character}. Type of legend to display. One of: normal, colorbar, colorsteps.}

\item{legend.position}{\link[base]{character}. Position of the legend in the plot. One of: top, bottom, left, right or none (if no legend is desired).}

\item{legend.framewidth, legend.tickwidth}{\link[base]{numeric}. Width of the lines of the box in the legend.}

\item{legend.framecolor, legend.tickcolor}{\link[base]{character}. Color of the lines of the box in the legend.}

\item{legend.length, legend.width}{\link[base]{numeric}. Length and width of the legend. Will adjust automatically depending on legend side.}

\item{plot.title, plot.subtitle, plot.caption}{\link[base]{character}. Title, subtitle or caption to use in the plot.}

\item{assay}{\link[base]{character}. Assay to use. Defaults to the current assay.}

\item{slot}{\link[base]{character}. Data slot to use. \link[base]{character}. Only one of: counts, data, scale.data. Defaults to "data".}

\item{reduction}{\link[base]{character}. Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}. Defaults to "umap" if present or to the last computed reduction if the argument is not provided.}

\item{viridis_color_map}{\link[base]{character}. A capital letter from A to H or the scale name as in \link[viridis]{scale_fill_viridis}.}

\item{raster}{\link[base]{logical}. Whether to raster the resulting plot. This is recommendable if plotting a lot of cells.}

\item{raster.dpi}{\link[base]{numeric}. Pixel resolution for rasterized plots. Defaults to 512, as per default \code{Seurat::FeaturePlot()} behavior.}

\item{plot_cell_borders}{\link[base]{logical}. Whether to plot border around cells.}

\item{border.size}{\link[base]{numeric}. Width of the border of the cells.}

\item{border.color}{\link[base]{character}. Color to use for the border of the cells.}

\item{na.value}{\link[base]{character}. Color value for NA.}

\item{rotate_x_axis_labels}{\link[base]{logical}. Whether to rotate X axis labels.}

\item{xlab, ylab}{\link[base]{character}. Titles for the X and Y axis.}
}
\value{
Nothing. This is a mock function.
}
\description{
Mock function used to document all main function.
}
\examples{
#' \dontrun{
TBD
}
}
