% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in operator.}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{First item of the comparison.}

\item{y}{Second item of the comparison.}
}
\value{
A logical vector of the items in x that are not in y.
}
\description{
Not in operator.
}
\examples{
\dontrun{
TBD
}
}
