% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_chr_locations}
\alias{human_chr_locations}
\title{Chromosome arm locations for human genome GRCh38.}
\format{
A tibble with 48 rows and 4 columns:
\describe{
\item{chr}{Chromosome.}
\item{arm}{Chromosome arm.}
\item{start}{Start coordinates.}
\item{end}{End coordinates.}
}
}
\usage{
data(human_chr_locations)
}
\description{
A tibble containing the chromosome, arm and start and end coordinates.
}
\keyword{datasets}
