% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modify_string}
\alias{modify_string}
\title{Modify a string to wrap it around the middle point.}
\usage{
modify_string(string_to_modify)
}
\arguments{
\item{string_to_modify}{}
}
\value{

}
\description{
Modify a string to wrap it around the middle point.
}
