% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_list}
\alias{named_list}
\title{Named list.}
\value{
Nothing. This is a mock function.
}
\description{
Named list.
}
\examples{
# This is a named vector.
x <- list("first_element" = c("GENE A", "GENE B"),
          "second_element" = c("GENE C", "GENE D"))
print(x)

}
\keyword{internal}
