% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_vector}
\alias{named_vector}
\title{Named vector.}
\value{
Nothing. This is a mock function.
}
\description{
Named vector.
}
\examples{
# This is a named vector.
x <- c("first_element" = 3,
       "second_element" = TRUE)
print(x)

}
\keyword{internal}
