% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_Plot.R
\name{save_Plot}
\alias{save_Plot}
\title{Save a plot as png, pdf and svg.}
\usage{
save_Plot(
  plot,
  figure_path = NULL,
  create_path = TRUE,
  file_name = NULL,
  dpi = 300,
  output_format = "publication",
  width = 8,
  height = 8
)
}
\arguments{
\item{plot}{Plot to save.}

\item{figure_path}{Path where the figure will be stored.}

\item{create_path}{Logical. Whether to create the path.}

\item{file_name}{Name of the file (without extension, it will be added automatically).}

\item{dpi}{Dpi to use.}

\item{output_format}{Character. One of the following:
\itemize{
\item pdf.
\item png.
\item jpeg.
\item tiff.
\item svg
\item publication: Includes pdf, png and svg formats.
\item all: Includes all formats.
}}

\item{width, height}{Width and height of the figure (inches).}
}
\value{
Nothing.
}
\description{
Save a plot as png, pdf and svg.
}
\examples{
\donttest{
  # Generate a plot.
  p <- SCpubr::do_DimPlot(sample = sample)

  # Default parameters.
  SCpubr::savePlot(plot = p)

  # Specifying the name and folder.
  SCpubr::savePlot(plot = p,
                   figure_path = "/path/to/my/figures/",
                   file_name = "my_figure")

  # Specify to also create a new folder.
  SCpubr::savePlot(plot = p,
                   figure_path = "/path/to/my/figures/",
                   file_name = "my_figure",
                   create_path = TRUE)

  # Set dimensions for the figure.
  SCpubr::savePlot(plot = p,
                   figure_path = "/path/to/my/figures/",
                   file_name = "my_figure",
                   create_path = TRUE,
                   width = 8,
                   height = 8)

  # Set quality (dpi).
  SCpubr::savePlot(plot = p,
                   figure_path = "/path/to/my/figures/",
                   file_name = "my_figure",
                   create_path = TRUE,
                   width = 8,
                   height = 8,
                   dpi = 300)
}
}
