% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePlot.R
\name{savePlot}
\alias{savePlot}
\title{Save a plot as png, pdf and svg.}
\usage{
savePlot(
  plot,
  figure_path = NULL,
  create_path = TRUE,
  file_name = NULL,
  dpi = 300,
  output_format = "publication",
  width = 8,
  height = 8
)
}
\arguments{
\item{plot}{Plot to save.}

\item{figure_path}{Path where the figure will be stored.}

\item{file_name}{Name of the file (without extension, it will be added automatically).}

\item{dpi}{Dpi to use.}

\item{output_format}{Character. One of the following:
\itemize{
\item pdf.
\item png.
\item jpeg.
\item tiff.
\item svg
\item publication: Includes pdf, png and svg formats.
\item all: Includes all formats.
}}

\item{width, height}{Width and height of the figure (inches).}

\item{create_folder}{Logical. Whether to create the path.}
}
\value{

}
\description{
Save a plot as png, pdf and svg.
}
\examples{
\dontrun{
TBD
}
}
