% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_BarPlot.R
\name{do_BarPlot}
\alias{do_BarPlot}
\title{Wrapper for computing publication ready bar plots.}
\usage{
do_BarPlot(
  sample,
  var.to.plot,
  group.by = NULL,
  labels.order = NULL,
  order.by = NULL,
  position = "stack",
  xlab = "",
  ylab = "",
  plot.title = "",
  legend.position = "bottom",
  legend.title = FALSE,
  legend.ncol = 1,
  legend.text.size = 22,
  legend.title.size = 22,
  axis.text.size = 22,
  axis.title.size = 22,
  plot.title.size = 24,
  legend.icon.size = 4,
  legend.byrow = FALSE,
  colors.use = NULL,
  horizontal = TRUE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{var.to.plot}{Main variable in the bar plot. Example: seurat_clusters}

\item{group.by}{Secondary variable to group the bar plot for. Example: orig.ident}

\item{labels.order}{Vector of labels to explicitly state the order or the bars.}

\item{order.by}{Value in var.to.plot to order the items in group.by for.}

\item{position}{Either "fill" or "stack." Posigion "fill" will generate a bar plot with one column and the proportions of values for each group inside, while "stack" plots the bars together.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{plot.title}{Title to use in the plot.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.title}{Logical stating whether the legend title is shown or not.}

\item{legend.text.size}{Fontsize of the legend labels.}

\item{legend.title.size}{Fontisize of the legend title.}

\item{axis.text.size}{Fontsize for axis text.}

\item{axis.title.size}{Fontsize for axis title.}

\item{plot.title.size}{Fontsize for the plot title.}

\item{legend.icon.size}{Size of the icons in legend.}

\item{legend.byrow}{Logical stating whether the legend is filled by row or not.}

\item{colors.use}{Palette of colors to use. It must match the group.by variable in terms of length and names.}

\item{legend.col}{Number of columns in the legend.}
}
\value{

}
\description{
Wrapper for computing publication ready bar plots.
}
