% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_BarPlot.R
\name{do_BarPlot}
\alias{do_BarPlot}
\title{Wrapper for computing publication ready bar plots.}
\usage{
do_BarPlot(
  sample,
  features,
  group.by = NULL,
  labels.order = NULL,
  order.by = NULL,
  position = "stack",
  xlab = NULL,
  ylab = NULL,
  plot.title = NULL,
  legend = TRUE,
  legend.position = "right",
  legend.title = FALSE,
  legend.ncol = 1,
  fontsize = 14,
  legend.icon.size = 4,
  legend.byrow = FALSE,
  colors.use = NULL,
  horizontal = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Main variable in the bar plot - This will define the height of the bars. Must be a metadata variable stored in \code{object@meta.data}.}

\item{group.by}{Secondary variable to group the bar plot for - This will define the total number of bars. Must be a metadata variable stored in \code{object@meta.data}.}

\item{labels.order}{Vector of labels to explicitly state the order or the bars. Better fed after knowing the order. This is meant to be used to tweak the end plot if needed.}

\item{order.by}{One of the unique values in group.by that will be used to reorder the bars according to its descending proportion.}

\item{position}{Either "fill" or "stack." Position "fill" will generate a bar plot with one column and the proportions of values for each group inside, while "stack" plots the bars together.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{plot.title}{Title to use in the plot.}

\item{legend}{Whether to plot the legend.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.title}{Logical stating whether the legend title is shown or not.}

\item{legend.ncol}{Number of columns in the legend.}

\item{fontsize}{Base fontsize of the figure.}

\item{legend.icon.size}{Size of the icons in legend.}

\item{legend.byrow}{Logical stating whether the legend is filled by row or not.}

\item{colors.use}{Palette of colors to use. It must match the group.by variable in terms of length and names.}

\item{horizontal}{Whether to plot the Bar plot horizontally.}

\item{verbose}{Use warnings.}
}
\value{
A ggplot2 object containing a Bar plot.
}
\description{
Wrapper for computing publication ready bar plots.
}
\examples{
\dontrun{
TBD
}
}
