% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_BeeSwarmPlot.R
\name{do_BeeSwarmPlot}
\alias{do_BeeSwarmPlot}
\title{BeeSwarm plot.}
\usage{
do_BeeSwarmPlot(
  sample,
  feature_to_rank,
  group.by,
  assay = NULL,
  reduction = NULL,
  slot = NULL,
  continuous_feature = FALSE,
  colors.use = NULL,
  legend.position = "right",
  plot.title = "",
  xlab = NULL,
  ylab = "",
  fontsize = 14,
  remove_x_axis = FALSE,
  remove_y_axis = FALSE,
  flip = FALSE,
  viridis_color_map = "D",
  verbose = TRUE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{feature_to_rank}{Features for which the cells are going to be ranked. Ideal case is that this feature is stored as a metadata column.}

\item{group.by}{Variable you want the cells to be grouped for.}

\item{assay}{Assay to use. Defauls to the current assay.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}. Defaults to "umap" if present or to the last computed reduction if the argument is not provided.}

\item{slot}{Data slot to use. Character. Only one of: counts, data, scale.data. Defaults to "data".}

\item{continuous_feature}{Is the feature to rank and color for continuous? I.e: an enrichment score.}

\item{colors.use}{Named vector with the color assignment.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{plot.title}{Title to use in the plot.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{fontsize}{Base fontsize of the plot.}

\item{remove_x_axis}{Remove X axis labels and ticks from the plot.}

\item{remove_y_axis}{Remove Y axis labels and ticks from the plot.}

\item{flip}{Whether to flip the axis.}

\item{viridis_color_map}{Character. A capital letter from A to H or the scale name as in \link[viridis]{scale_fill_viridis}.}

\item{verbose}{Whether to show warnings.}

\item{legend}{Whether to plot the legend or not.}
}
\value{
A ggplot2 object containing a Bee Swarm plot.
}
\description{
BeeSwarm plot.
}
\examples{
\dontrun{
TBD
}
}
