% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_BeeSwarmPlot.R
\name{do_BeeSwarmPlot}
\alias{do_BeeSwarmPlot}
\title{BeeSwarm plot.}
\usage{
do_BeeSwarmPlot(
  sample,
  feature_to_rank,
  group.by,
  assay = NULL,
  reduction = NULL,
  slot = NULL,
  continuous_feature = FALSE,
  colors.use = NULL,
  legend.type = "colorbar",
  legend.position = "bottom",
  legend.framewidth = 1.5,
  legend.tickwidth = 1.5,
  legend.length = 20,
  legend.width = 1,
  legend.framecolor = "grey50",
  legend.tickcolor = "white",
  plot.title = NULL,
  plot.subtitle = NULL,
  plot.caption = NULL,
  xlab = NULL,
  ylab = "",
  font.size = 14,
  font.type = "sans",
  remove_x_axis = FALSE,
  remove_y_axis = FALSE,
  flip = FALSE,
  viridis_color_map = "D",
  verbose = TRUE,
  raster = FALSE,
  raster.dpi = 300,
  plot_cell_borders = TRUE,
  border.size = 1.5,
  border.color = "black",
  pt.size = 2
)
}
\arguments{
\item{sample}{Seurat object.}

\item{feature_to_rank}{Feature for which the cells are going to be ranked. Ideal case is that this feature is stored as a metadata column.}

\item{group.by}{Variable you want the cells to be grouped for.}

\item{assay}{Assay to use. Defauls to the current assay.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}. Defaults to "umap" if present or to the last computed reduction if the argument is not provided.}

\item{slot}{Data slot to use. Character. Only one of: counts, data, scale.data. Defaults to "data".}

\item{continuous_feature}{Is the feature to rank and color for continuous? I.e: an enrichment score.}

\item{colors.use}{Named vector with the color assignment.}

\item{legend.type}{Character. Type of legend to display. One of: normal, colorbar, colorsteps.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.framewidth, legend.tickwidth}{Width of the lines of the box in the legend.}

\item{legend.length, legend.width}{Length and width of the legend. Will adjust automatically depending on legend side.}

\item{legend.framecolor, legend.tickcolor}{Color of the lines of the box in the legend.}

\item{plot.title, plot.subtitle, plot.caption}{Title, subtitle or caption to use in the plot.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{font.size}{Base font.size of the plot.}

\item{font.type}{Character. Base font for the plot. One of mono, serif or sans.}

\item{remove_x_axis}{Remove X axis labels and ticks from the plot.}

\item{remove_y_axis}{Remove Y axis labels and ticks from the plot.}

\item{flip}{Whether to flip the axis.}

\item{viridis_color_map}{Character. A capital letter from A to H or the scale name as in \link[viridis]{scale_fill_viridis}.}

\item{verbose}{Whether to show warnings.}

\item{raster}{Whether to raster the resulting plot. This is recommendable if plotting a lot of cells.}

\item{raster.dpi}{Numeric. Resolution in dots per inch. Defaults to 5.}

\item{plot_cell_borders}{Logical. Whether to plot border around cells.}

\item{border.size}{Numeric. Width of the border of the cells.}

\item{border.color}{Character. Color to use for the border of the cells.}

\item{pt.size}{Numeric. Size of the dots.}

\item{legend}{Whether to plot the legend or not.}
}
\value{
A ggplot2 object containing a Bee Swarm plot.
}
\description{
BeeSwarm plot.
}
\examples{
\donttest{
  # Basic Bee Swarm plot - categorical coloring.
  # This will color based on the unique values of seurat_clusters.
  p <- SCpubr::do_BeeSwarmPlot(sample = sample,
                              feature_to_rank = "PC_1",
                              group.by = "seurat_clusters",
                              continuous_feature = F)

  # Basic Bee Swarm plot - continuous coloring.
  # This will color based on the PC_1 values.
  p <- SCpubr::do_BeeSwarmPlot(sample = sample,
                               feature_to_rank = "PC_1",
                               group.by = "seurat_clusters",
                               continuous_feature = T)

  # Change default colors - categorical variables.
  colors <- c("0" = "#001219",
              "1" = "#005f73",
              "2" = "#0a9396",
              "3" = "#94d2bd",
              "4" = "#e9d8a6",
              "5" = "#ee9b00",
              "6" = "#ca6702",
              "7" = "#bb3e03",
              "8" = "#ae2012",
              "9" = "#9b2226")

  p <- SCpubr::do_BeeSwarmPlot(sample = sample,
                               feature_to_rank = "Monocyte_signature",
                               group.by = "seurat_clusters",
                               colors.use = colors)

  # Change default colors - continuous variables.
  p <- SCpubr::do_BeeSwarmPlot(sample = sample,
                               feature_to_rank = "Monocyte_signature",
                               group.by = "seurat_clusters",
                               continuous_feature = TRUE,
                               fontsize = 10,
                               viridis_color_map = "A",
                               plot.title = "Magma")


}
}
