% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_BeeSwarmPlot.R
\name{do_BeeSwarmPlot}
\alias{do_BeeSwarmPlot}
\title{BeeSwarm plot.}
\usage{
do_BeeSwarmPlot(
  sample,
  feature_to_rank,
  group.by,
  assay = "SCT",
  reduction = "umap",
  continuous_feature = FALSE,
  colors.use = NULL,
  legend.position = "right",
  plot.title = "",
  xlab = NULL,
  ylab = "",
  axis.text.fonsize = 15,
  axis.title.fonsize = 15,
  plot.title.fonsize = 15,
  legend.text.fontsize = 10,
  legend.title.fontsize = 12,
  remove_x_axis = FALSE,
  remove_y_axis = FALSE,
  flip = FALSE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{feature_to_rank}{Features for which the cells are going to be ranked. Ideal case is that this feature is stored as a metadata column.}

\item{group.by}{Variable you want the cells to be grouped for.}

\item{assay}{Seurat assay to choose.}

\item{reduction}{Reduction associated to feature_to_rank if this feature is a component like PC_1, UMAP_1 or similar.}

\item{continuous_feature}{Is the feature to rank and color for continuous? I.e: an enrichment score.}

\item{colors.use}{Named vector with the color assignment.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{plot.title}{Title to use in the plot.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{axis.text.fonsize}{Modify the font size for axis texts.}

\item{axis.title.fonsize}{Modify the font size for axis titles.}

\item{plot.title.fonsize}{Modify the font size for the plot title.}

\item{legend.text.fontsize}{Modify the font size for the legend text.}

\item{legend.title.fontsize}{Modify the font size for the legend title.}

\item{remove_x_axis}{Remove X axis labels and ticks from the plot.}

\item{remove_y_axis}{Remove Y axis labels and ticks from the plot.}

\item{flip}{Whether to flip the axis.}

\item{legend}{Whether to plot the legend or not.}
}
\value{
A ggplot2 object containing a Bee Swarm plot.
}
\description{
BeeSwarm plot.
}
\examples{
\dontrun{
TBD
}
}
