% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_CellularStatesPlot.R
\name{do_CellularStatesPlot}
\alias{do_CellularStatesPlot}
\title{Butterfly plot.}
\usage{
do_CellularStatesPlot(
  sample,
  gene_list,
  x1,
  x2 = NULL,
  y1,
  y2 = NULL,
  group.by = NULL,
  colors.use = NULL,
  legend.position = NULL,
  plot.title = "",
  plot.subtitle = NULL,
  plot.caption = NULL,
  xlab = NULL,
  ylab = NULL,
  axis.ticks = TRUE,
  axis.text = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{gene_list}{Named list of lists of marker genes to query for enrichment.}

\item{x1}{A name of a list from gene_list. First feature in the X axis. Will go on the right side of the X axis if y2 is not provided and top-right quadrant if provided.}

\item{x2}{A name of a list from gene_list. Second feature on the X axis. Will go on the left side of the X axis if y2 is not provided and top-left quadrant if provided.}

\item{y1}{A name of a list from gene_list. First feature on the Y axis. Will become the Y axis if y2 is not provided and bottom-right quadrant if provided.}

\item{y2}{A name of a list from gene_list. Second feature on the Y axis. Will become the bottom-left quadrant if provided.}

\item{group.by}{Metadata variable to color the cells by. Defaults to current identities.}

\item{colors.use}{Named vector with the names of the unique values in the categorical variable and values the HEX codes for the colors.}

\item{legend.position}{Position of the legend in the plot. One of: top, bottom, left, right.}

\item{plot.title, plot.subtitle, plot.caption}{Title to use in the plot.}

\item{xlab}{Title for the X axis. Only works if y2 is not set up.}

\item{ylab}{Title for the Y axis. Only works if y2 is not set up.}

\item{axis.ticks}{Whether to show axis ticks.}

\item{axis.text}{Whether to show axis text.}

\item{verbose}{Verbose function?}
}
\value{
A ggplot2 object containing a butterfly plot.
}
\description{
This plot aims to show the relationships between distinct enrichment scores. If 3 variables are provided, the relationship is between the Y axis and the dual X axis.
If 4 variables are provided, each corner of the plot represents how enriched the cells are in that given list. How to interpret this? In a 3-variable plot, the Y axis
just means one variable. The higher the cells are in the Y axis the more enriched they are in that given variable. The X axis is a dual parameter one. Cells falling
into each extreme of the axis are highly enriched for either x1 or x2, while cells falling in between are not enriched for any of the two. In a 4-variable plot, each corner
shows the enrichment for one of the 4 given features. Cells will tend to locate in either of the four corners, but there will be cases of cells locating mid-way between two
given corners (enriched in both features) or in the middle of the plot (not enriched for any).
}
\details{
This plots are based on the following publications:
\itemize{
\item Neftel, C. et al. An Integrative Model of Cellular States, Plasticity, and Genetics for Glioblastoma. Cell 178, 835-849.e21 (2019). https://doi.org/10.1016/j.cell.2019.06.024
\item Tirosh, I., Venteicher, A., Hebert, C. et al. Single-cell RNA-seq supports a developmental hierarchy in human oligodendroglioma. Nature 539, 309–313 (2016). https://doi.org/10.1038/nature20123
}
}
\examples{
\dontrun{
TBD
}
}
