% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_CorrelationPlot.R
\name{do_CorrelationPlot}
\alias{do_CorrelationPlot}
\title{Create correlation matrix heatmaps.}
\usage{
do_CorrelationPlot(
  sample,
  mode = "hvg",
  assay = NULL,
  group.by = NULL,
  column_title = NULL,
  row_title = NULL,
  cluster_cols = TRUE,
  cluster_rows = TRUE,
  legend_name = "Pearson coef.",
  row_names_rot = 0,
  column_names_rot = 90,
  colors.use = NULL,
  cell_size = 5,
  na.value = "grey75"
)
}
\arguments{
\item{sample}{Your Seurat object.}

\item{mode}{Different types of correlation matrices can be computed. Right now, the only possible value is "hvg", standing for Highly Variable Genes. The sample is subset for the HVG and the data is re-scaled. Scale data is used for the correlation.}

\item{assay}{Assay to retrieve the data from.}

\item{group.by}{Metadata variable to group the values by.}

\item{column_title, row_title}{Title for the columns and rows. Only works with single heatmaps.}

\item{cluster_cols, cluster_rows}{Logical. Cluster the columns or rows.}

\item{legend_name}{Text for the legend title.}

\item{column_names_rot, row_names_rot}{Numeric. Degree in which to}

\item{colors.use}{Vector of 2 colors to use to generate the color scale.}

\item{cell_size}{Numeric. Size of each cell in the heatmap.}

\item{na.value}{Value for NAs.}
}
\value{
A ComplexHeatmap object.
}
\description{
Create correlation matrix heatmaps.
}
\examples{
\donttest{
  # Default values.
  p <- SCpubr::do_CorrelationPlot(sample = sample)
  p

  # Custom grouping.
  clusters <- c("1", "3", "5", "7", "9")
  sample$custom_group <- ifelse(sample$seurat_clusters \%in\% clusters,
                                "Group A",
                                "Group B")
  p <- SCpubr::do_CorrelationPlot(sample = sample, group.by = "custom_group")
  p

  # Rotated axis labels.
  p <- SCpubr::do_CorrelationPlot(sample = sample,
                                  column_names_rot = 0)
  p

  # Increase cell size.
  p <- SCpubr::do_CorrelationPlot(sample = sample,
                                  column_names_rot = 0,
                                  cell_size = 7)
  p

  # Modified color scale.
  p <- SCpubr::do_CorrelationPlot(sample = sample,
                                  column_names_rot = 0,
                                  colors.use = colortools::opposite("steelblue", plot = F))
  p
}
}
