% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_DimPlot.R
\name{do_DimPlot}
\alias{do_DimPlot}
\title{Wrapper for \link[Seurat]{DimPlot}.}
\usage{
do_DimPlot(
  sample,
  reduction = NULL,
  group.by = NULL,
  split.by = NULL,
  colors.use = NULL,
  shuffle = TRUE,
  order = NULL,
  pt.size = 1,
  label = FALSE,
  label.color = "white",
  repel = TRUE,
  cells.highlight = NULL,
  idents.highlight = NULL,
  idents.keep = NULL,
  sizes.highlight = 1,
  legend = TRUE,
  ncol = NULL,
  plot.title = NULL,
  plot.subtitle = NULL,
  plot.caption = NULL,
  legend.title = FALSE,
  legend.position = "bottom",
  legend.title.position = "top",
  legend.ncol = NULL,
  legend.nrow = NULL,
  legend.icon.size = 4,
  legend.byrow = FALSE,
  raster = FALSE,
  raster.dpi = 1024,
  dims = c(1, 2),
  font.size = 14,
  font.type = "sans",
  na.value = "grey75",
  plot_cell_borders = TRUE,
  border.size = 2,
  border.color = "black",
  plot_marginal_distributions = FALSE,
  marginal.type = "density",
  marginal.size = 5,
  marginal.group = TRUE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}. Defaults to "umap" if present or to the last computed reduction if the argument is not provided.}

\item{group.by}{Variable you want the cells to be colored for.}

\item{split.by}{Split into as many plots as unique values in the variable provided.}

\item{colors.use}{Vector of named HEX values to color the cells. It has to match the number of unique values in either \code{Seurat::Idents(sample)} or the group.by or split.by variable. For split.by, a single color can be provided and each panel will be colored by it.}

\item{shuffle}{Whether to shuffle the cells or not, so that they are not plotted cluster-wise. Recommended.}

\item{order}{Vector of identities to be plotted. Either one with all identities or just some, which will be plotted last.}

\item{pt.size}{Point size of the cells.}

\item{label}{Whether to plot the cluster labels in the UMAP. The cluster labels will have the same color as the cluster colors.}

\item{label.color}{HEX code for the color of the text in the labels if label is set to TRUE.}

\item{repel}{Whether to repel the labels if label is set to TRUE.}

\item{cells.highlight, idents.highlight}{Vector of cells/identities for which the DimPlot should focus into. The rest of the cells will be grayed out. Both parameters can be used at the same time.}

\item{idents.keep}{Vector of identities to keep. This will effectively set the rest of the cells that do not match the identities provided to NA, therefore coloring them according to na.value parameter.}

\item{sizes.highlight}{Point size of highlighted cells using cells.highlight parameter.}

\item{legend}{Whether to plot the legend or not.}

\item{ncol}{Number of columns used in the arrangement of the output plot using "split.by" parameter.}

\item{plot.title, plot.subtitle, plot.caption}{Title, subtitle or caption to use in the plot.}

\item{legend.title}{Logical stating whether the legend title is shown or not.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.title.position}{Character stating where to place the title of the legend.}

\item{legend.ncol, legend.nrow}{Number of columns/rows in the legend.}

\item{legend.icon.size}{Numeric. Size of the icons in legend.}

\item{legend.byrow}{Logical stating whether the legend is filled by row or not.}

\item{raster}{Whether to raster the resulting plot. This is recommendable if plotting a lot of cells.}

\item{raster.dpi}{Numeric. Pixel resolution for rasterized plots. Defaults to 512, as per default \code{Seurat::DimPlot()} behavior.}

\item{dims}{Vector of 2 numerics indicating the dimensions to plot out of the selected reduction. Defaults to c(1, 2) if not specified.}

\item{font.size}{Base font.size of the figure.}

\item{font.type}{Character. Base font for the plot. One of mono, serif or sans.}

\item{na.value}{Color value for NA.}

\item{plot_cell_borders}{Logical. Whether to plot border around cells.}

\item{border.size}{Numeric. Width of the border of the cells.}

\item{border.color}{Character. Color to use for the border of the cells.}

\item{plot_marginal_distributions}{Logical. Whether to plot marginal distributions on the figure or not.}

\item{marginal.type}{Character. One of density", "histogram", "boxplot", "violin", "densigram". Defaults to density.}

\item{marginal.size}{Numeric. Size ratio between the main and marginal plots. 5 means that the main plot is 5 times bigger than the marginal plots.}

\item{marginal.group}{Logical. Whether to group the marginal distribution by group.by or current identities.}
}
\value{
A ggplot2 object containing a DimPlot.
}
\description{
Wrapper for \link[Seurat]{DimPlot}.
}
\examples{
\donttest{
# Your Seurat object.
sample <- my_seurat_object

# Basic DimPlot.
p <- SCpubr::do_DimPlot(sample = sample)

# Control dimensions.
p <- SCpubr::do_DimPlot(sample = sample,
                        plot.title = "My awesome SC data set",
                        dims = c(2, 1))

# Include a plot title.
p <- SCpubr::do_DimPlot(sample = sample,
                        plot.title = "My awesome SC data set")

# Include a plot subtitle.
p <- SCpubr::do_DimPlot(sample = sample,
                        plot.subtitle = "My awesome SC data set")

# Include a plot caption
p <- SCpubr::do_DimPlot(sample = sample,
                        plot.caption = "My awesome SC data set")

# Control legend position and number of columns.
p <- SCpubr::do_DimPlot(sample = sample,
                        plot.title = "My awesome SC data set",
                        legend.position = "left",
                        legend.ncol = 2)

# Control legend position and number of rows
p <- SCpubr::do_DimPlot(sample = sample,
                        plot.title = "My awesome SC data set",
                        legend.position = "left",
                        legend.nrow = 2)

# Use labels instead of legend.
p <- SCpubr::do_DimPlot(sample = sample,
                        label = TRUE,
                        legend = FALSE)

# Changing the order of plotting.
# Using order with one identity value.
p <- SCpubr::do_DimPlot(sample = sample,
                        shuffle = TRUE,
                        order = "5")
# Using order with all identity values.
p <- SCpubr::do_DimPlot(sample = sample,
                        shuffle = FALSE,
                        order = c("5", "8", "4",
                                  "9", "3", "1",
                                  "6", "0", "7", "2"))

# Restrict the amount of identities displayed.
p <- SCpubr::do_DimPlot(sample = sample,
                        idents.keep = c("1", "3", "5"))


# Group by another variable rather than `Seurat::Idents(sample)`
p <- SCpubr::do_DimPlot(sample = sample,
                        group.by = "seurat_clusters")

# Split the output in as many plots as unique identities.
p <- SCpubr::do_DimPlot(sample = sample,
                        split.by = "seurat_clusters")

# Restrict the amount of identities displayed by split.by.
p <- SCpubr::do_DimPlot(sample = sample,
                        split.by = "seurat_clusters",
                        idents.keep = c("1", "3", "5"))

# Modify default colors.
# Colors need to a named vector of the same length as the identities.
colors <- c("0" = "#001219",
            "1" = "#005f73",
            "2" = "#0a9396",
            "3" = "#94d2bd",
            "4" = "#e9d8a6",
            "5" = "#ee9b00",
            "6" = "#ca6702",
            "7" = "#bb3e03",
            "8" = "#ae2012",
            "9" = "#9b2226")
p <- SCpubr::do_DimPlot(sample = sample,
                        colors.use = colors)

# Highlight cells.
# Select 1000 random cells out of clusters 1, 5 and 7.
cells.use <- sample(colnames(sample[, sample$seurat_clusters \%in\% c("1", "5", "7")]), 1000)
p <- SCpubr::do_DimPlot(sample = sample,
                        cells.highlight = cells.use)

# Highlight cells and use custom color.
# Select 1000 random cells out of clusters 1, 5 and 7.
cells.use <- sample(colnames(sample[, sample$seurat_clusters \%in\% c("1", "5", "7")]), 1000)
p <- SCpubr::do_DimPlot(sample = sample,
                        cells.highlight = cells.use,
                        colors.use = "black")

# Change the size of the highlighted cells.
# Select 1000 random cells out of clusters 1, 5 and 7.
cells.use <- sample(colnames(sample[, sample$seurat_clusters \%in\% c("1", "5", "7")]), 1000)
p <- SCpubr::do_DimPlot(sample = sample,
                        cells.highlight = cells.use,
                        colors.use = "black",
                        sizes.highlight = 1)

# Highlight given identities
p <- Seurat::DimPlot(sample,
                     idents.highlight = c("1", "3"))

# Highlight given cells and given identities.
cells.use <- sample(colnames(sample[, sample$seurat_clusters \%in\% c("1", "5", "7")]), 1000)
p <- Seurat::DimPlot(sample,
                     cells.highlight = cells.use,
                     idents.highlight = c("2", "4"))

}


}
