% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_DimPlot.R
\name{do_DimPlot}
\alias{do_DimPlot}
\title{Wrapper for \link[Seurat]{DimPlot}.}
\usage{
do_DimPlot(
  sample,
  reduction = "umap",
  label = FALSE,
  label.color = "black",
  repel = TRUE,
  shuffle = TRUE,
  pt.size = 0.5,
  sizes.highlight = 0.5,
  group.by = NULL,
  split.by = NULL,
  colors.split = "#0A305F",
  cells.highlight = NULL,
  colors.highlight = "#0A305F",
  legend = TRUE,
  legend.title = FALSE,
  legend.position = "bottom",
  legend.ncol = 4,
  legend.text.size = 14,
  legend.title.size = 14,
  legend.icon.size = 4,
  legend.byrow = FALSE,
  colors.use = NULL,
  plot.title = "",
  ncol = NULL,
  raster = FALSE,
  dims = c(1, 2)
)
}
\arguments{
\item{sample}{Seurat object.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}.}

\item{label}{Whether to plot the cluster labels in the UMAP. The cluster labels will have the same color as the cluster colors.}

\item{label.color}{HEX code for the color of the text in the labels if label is set to TRUE.}

\item{repel}{Whether to repel the labels if label is set to TRUE.}

\item{shuffle}{Whether to shuffle the cells or not, so that they are not plotted cluster-wise. Recommended.}

\item{pt.size}{Point size of the cells.}

\item{sizes.highlight}{Point size of highlighted cells using cells.highlight parameter.}

\item{group.by}{Variable you want the cells to be colored for.}

\item{split.by}{Split into as many plots as unique values in the variable provided.}

\item{colors.split}{Vector of named HEX values to color the cells in a split DimPlot. It has to match the unique values in split.by parameter.}

\item{cells.highlight}{Vector of cells for which the DimPlot should focus into. The rest of the cells will be grayed out.}

\item{colors.highlight}{HEX color code to use with the highlighted cells.}

\item{legend}{Whether to plot the legend or not.}

\item{legend.title}{Logical stating whether the legend title is shown or not.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.ncol}{Number of columns in the legend.}

\item{legend.text.size}{Font size of the legend labels.}

\item{legend.title.size}{Font size of the legend title.}

\item{legend.icon.size}{Size of the icons in legend.}

\item{legend.byrow}{Logical stating whether the legend is filled by row or not.}

\item{colors.use}{Vector of named HEX values to color the cells. It has to match the number of unique values in either \code{Seurat::Idents(sample)} or the group.by variable.}

\item{plot.title}{Title to use in the plot.}

\item{ncol}{Number of columns used in the arrangement of the output plot using "split.by" parameter.}

\item{raster}{Whether to raster the resulting plot. This is recommendable if plotting a lot of cells.}

\item{dims}{Dimensions to plot, if the dimensional reduction chosen has more than 2, like PCA.}
}
\value{
A ggplot2 object containing a DimPlot.
}
\description{
Wrapper for \link[Seurat]{DimPlot}.
}
\examples{
\dontrun{
TBD
}
}
