% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_DotPlot.R
\name{do_DotPlot}
\alias{do_DotPlot}
\title{Wrapper for \link[Seurat]{DotPlot}.}
\usage{
do_DotPlot(
  sample,
  features,
  assay = NULL,
  group.by = NULL,
  split.by = NULL,
  legend = TRUE,
  dot.scale = 6,
  cols = c("grey75", "#014f86"),
  legend.position = "right",
  plot.title = NULL,
  xlab = NULL,
  ylab = NULL,
  fontsize = 14,
  cluster.idents = FALSE,
  flip = FALSE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Features to represent.}

\item{assay}{Assay to use. Defaults to the current assay.}

\item{group.by}{Variable you want the cells to be colored for.}

\item{split.by}{Split into as many plots as unique values in the variable provided.}

\item{legend}{Whether to plot the legend or not.}

\item{dot.scale}{Scale the size of the dots.}

\item{cols}{Two colors if split.by is not set, which will define a gradient. As many numbers as unique values in split.by, if set, which each own will define its own gradient. Defaults to predefined color scales if not provided.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{plot.title}{Title to use in the plot.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{fontsize}{Base fontsize of the plot.}

\item{cluster.idents}{Logical. Whether to cluster the identities based on the expression of the features.}

\item{flip}{Whether to flip the axis.}
}
\value{
A ggplot2 object containing a Dot Plot.
}
\description{
Wrapper for \link[Seurat]{DotPlot}.
}
\examples{
\dontrun{
TBD
}
}
