% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_DotPlot.R
\name{do_DotPlot}
\alias{do_DotPlot}
\title{Wrapper for \link[Seurat]{DotPlot}.}
\usage{
do_DotPlot(
  sample,
  features,
  group.by = NULL,
  split.by = NULL,
  legend = TRUE,
  dot.scale = 10,
  cols = c("lightgrey", "#014f86"),
  legend.position = "bottom",
  plot.title = "",
  xlab = "",
  ylab = "",
  axis.text.fonsize = 28,
  axis.title.fonsize = 28,
  plot.title.fonsize = 30,
  legend.text.fontsize = 20,
  legend.title.fontsize = 24,
  remove_x_axis = FALSE,
  remove_y_axis = FALSE,
  flip = FALSE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Features to represent.}

\item{group.by}{Variable you want the cells to be colored for.}

\item{split.by}{Split into as many plots as unique values in the variable provided.}

\item{legend}{Whether to plot the legend or not.}

\item{dot.scale}{Scale the size of the dots.}

\item{cols}{From \link[Seurat]{DotPlot}: Colors to plot: the name of a palette from \code{RColorBrewer::brewer.pal.info}, a pair of colors defining a gradient, or 3+ colors defining multiple gradients (if split.by is set).}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{plot.title}{Title to use in the plot.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{axis.text.fonsize}{Modify the font size for axis texts.}

\item{axis.title.fonsize}{Modify the font size for axis titles.}

\item{plot.title.fonsize}{Modify the font size for the plot title.}

\item{legend.text.fontsize}{Modify the font size for the legend text.}

\item{legend.title.fontsize}{Modify the font size for the legend title.}

\item{remove_x_axis}{Remove X axis labels and ticks from the plot.}

\item{remove_y_axis}{Remove Y axis labels and ticks from the plot.}

\item{flip}{Whether to flip the axis.}
}
\value{
A ggplot2 object containing a Dot Plot.
}
\description{
Wrapper for \link[Seurat]{DotPlot}.
}
\examples{
\dontrun{
TBD
}
}
