% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_EnrichmentHeatmap.R
\name{do_EnrichmentHeatmap}
\alias{do_EnrichmentHeatmap}
\title{Create enrichment scores heatmaps.}
\usage{
do_EnrichmentHeatmap(
  sample,
  list_genes,
  group.by = NULL,
  column_title = "List of Genes",
  row_title = "Groups",
  verbose = FALSE,
  transpose = FALSE,
  cluster_cols = FALSE,
  cluster_rows = FALSE,
  legend_name = "Enrichment",
  row_names_rot = 0,
  column_names_rot = 90,
  colors.use = NULL,
  cell_size = 5,
  na.value = "grey75",
  legend.position = "bottom",
  use_viridis = TRUE,
  viridis_color_map = "G",
  scale_direction = -1,
  heatmap.legend.length = 75,
  heatmap.legend.width = 5,
  heatmap.legend.framecolor = "black",
  symmetrical_scale = FALSE,
  heatmap_gap = 0.5,
  row_names_side = "right",
  row_title_side = "left",
  row_title_rotation = 90
)
}
\arguments{
\item{sample}{Your Seurat object.}

\item{list_genes}{Named list of genes to query to the Seurat object.}

\item{group.by}{Metadata variable to group the values by.}

\item{column_title, row_title}{Title for the columns and rows. Only works with single heatmaps.}

\item{verbose}{Extended output.}

\item{transpose}{Logical. Transpose the resulting heatmap.}

\item{cluster_cols, cluster_rows}{Logical. Cluster the columns or rows.}

\item{legend_name}{Text for the legend title.}

\item{column_names_rot, row_names_rot}{Numeric. Degree in which to}

\item{colors.use}{Vector of 2 colors to use to generate the color scale.}

\item{cell_size}{Numeric. Size of each cell in the heatmap.}

\item{na.value}{Value for NAs.}

\item{legend.position}{Character. Location of the legend.}

\item{use_viridis}{Logical. Whether to use viridis color scales or a bicolor (blue-white-red) color scale.}

\item{viridis_color_map}{Character. Viridis color map for the heatmap. One of A, B, C, D, E, F, G, H.}

\item{scale_direction}{Numeric. Direction of the viridis scales. Either -1 or 1.}

\item{heatmap.legend.length, heatmap.legend.width}{Numeric. Width and length of the legend in the heatmap.}

\item{heatmap.legend.framecolor}{Character. Color of the edges and ticks of the legend in the heatmap.}

\item{symmetrical_scale}{Logical. Whether to make the scale symmetrical. Works best when use_viridis = FALSE.}

\item{heatmap_gap}{Numeric. Gap in cm between legends or heatmaps.}

\item{row_names_side}{Character. Either left or right.}

\item{row_title_side}{Character. Either left or right.}

\item{row_title_rotation}{Numeric. Degree of rotation of the row titles.}
}
\value{
A ComplexHeatmap object.
}
\description{
This function computes the enrichment scores for the cells using \link[Seurat]{AddModuleScore} and then aggregates the scores by the metadata variables provided by the user and displays it as a heatmap, computed by \link[ComplexHeatmap]{Heatmap}.
}
\examples{
\dontrun{
  # Define list of genes.
  genes <- list("Naive CD4+ T" = c("IL7R", "CCR7"),
                "CD14+ Mono" = c("CD14", "LYZ"),
                "Memory CD4+" = c("S100A4"),
                "B" = c("MS4A1"),
                "CD8+ T" = c("CD8A"),
                "FCGR3A+ Mono" = c("FCGR3A", "MS4A7"),
                "NK" = c("GNLY", "NKG7"),
                "DC" = c("FCER1A", "CST3"),
                "Platelet" = c("PPBP"))

  # Default parameters.
  p <- SCpubr::do_EnrichmentHeatmap(sample = sample,
                                    list_genes = genes)
  p

  # Custom aggregated values.
  p <- SCpubr::do_EnrichmentHeatmap(sample = sample,
                                    list_genes = genes,
                                    group.by = "orig.ident")
  p

  # Transposing the matrix.
  p <- SCpubr::do_EnrichmentHeatmap(sample = sample,
                                    list_genes = genes,
                                    transpose = TRUE)
  p

  # Rotating the labels.
  p <- SCpubr::do_EnrichmentHeatmap(sample = sample,
                                    list_genes = genes,
                                    transpose = TRUE,
                                    column_names_rot = 0)
  p

  # Modifying the tile size.
  p <- SCpubr::do_EnrichmentHeatmap(sample = sample,
                                    list_genes = genes,
                                    transpose = TRUE,
                                    column_names_rot = 0,
                                    cluster_cols = F,
                                    cluster_rows = T,
                                    split.by = "custom_group",
                                    cell_size = 7)
  p


  # Symmetrical scale viriis.
  p <- SCpubr::do_EnrichmentHeatmap(sample = sample,
                                    list_genes = genes,
                                    transpose = TRUE,
                                    column_names_rot = 0,
                                    cluster_cols = F,
                                    cluster_rows = T,
                                    cell_size = 7,
                                    symmetrical_scale = TRUE)
  p


  # Modifying the symmetrical scale non viridis.
  p <- SCpubr::do_EnrichmentHeatmap(sample = sample,
                                    list_genes = genes,
                                    transpose = TRUE,
                                    column_names_rot = 0,
                                    cluster_cols = F,
                                    cluster_rows = T,
                                    cell_size = 7,
                                    symmetrical_scale = TRUE,
                                    use_viridis = FALSE)
  p
}
}
