% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_FeaturePlot.R
\name{do_FeaturePlot}
\alias{do_FeaturePlot}
\title{Wrapper for \link[Seurat]{FeaturePlot}.}
\usage{
do_FeaturePlot(
  sample,
  assay = "SCT",
  features,
  reduction = "umap",
  pt.size = 0.5,
  legend = TRUE,
  legend.position = "right",
  plot.title = NULL,
  ncol = NULL,
  scale.begin = NULL,
  scale.end = NULL,
  cells.highlight = NULL,
  idents.highlight = NULL,
  dims = c(1, 2)
)
}
\arguments{
\item{sample}{Seurat object.}

\item{assay}{Assay to use.}

\item{features}{Features to plot. It can be a single one or a vector of multiple features. Similar behavior as with \link[Seurat]{FeaturePlot}.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use Seurat::Reductions(sample).}

\item{pt.size}{Point size.}

\item{legend}{Whether to plot the legend or not.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{plot.title}{Title for the plot.}

\item{ncol}{Number of columns to use in the arrangement of the output if more than one feature is queried to the function.}

\item{scale.begin}{Value to where you want the continuous color scale to start. It has to be in the range of the feature's values.}

\item{scale.end}{Value to where you want the continuous color scale to end. It has to be in the range of the feature's values.}

\item{cells.highlight}{Vector of cells for which the FeaturePlot should focus into. The rest of the cells will be grayed out.}

\item{idents.highlight}{Vector of identities that the FeaturePlot should focus into. Has to match the current Seurat identities in Seurat::Idents(sample).}

\item{dims}{Vector of 2 dimensions to use. Defaults to first and second dimensions.}
}
\value{

}
\description{
Wrapper for \link[Seurat]{FeaturePlot}.
}
