% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_FeaturePlot.R
\name{do_FeaturePlot}
\alias{do_FeaturePlot}
\title{Wrapper for \link[Seurat]{FeaturePlot}.}
\usage{
do_FeaturePlot(
  sample,
  assay = NULL,
  features,
  reduction = NULL,
  pt.size = 0.5,
  legend = TRUE,
  split.by = NULL,
  split.by.idents = NULL,
  slot = NULL,
  legend.position = "right",
  plot.title = NULL,
  fontsize = 14,
  ncol = NULL,
  cells.highlight = NULL,
  idents.highlight = NULL,
  dims = c(1, 2),
  viridis_color_map = "D",
  verbose = TRUE,
  individual.titles = NULL
)
}
\arguments{
\item{sample}{Seurat object.}

\item{assay}{Assay to use. Defauls to the current assay.}

\item{features}{Features to plot. It can be a single one or a vector of multiple features. Similar behavior as with \link[Seurat]{FeaturePlot}.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}. Defaults to "umap" if present or to the last computed reduction if the argument is not provided.}

\item{pt.size}{Point size.}

\item{legend}{Whether to plot the legend or not.}

\item{split.by}{Split the plot in as many unique values stored in the provided metadata column.}

\item{split.by.idents}{Vector of identities to plot. The gradient scale will also be subset to only the values of such identities.}

\item{slot}{Data slot to use. Character. Only one of: counts, data, scale.data. Defaults to "data".}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{plot.title}{Title for the plot.}

\item{fontsize}{Base fontsize of the figure.}

\item{ncol}{Number of columns to use in the arrangement of the output if more than one feature is queried to the function.}

\item{cells.highlight}{Vector of cells for which the FeaturePlot should focus into. The rest of the cells will be grayed out.}

\item{idents.highlight}{Vector of identities that the FeaturePlot should focus into. Has to match the current Seurat identities in \code{Seurat::Idents(sample)}.}

\item{dims}{Vector of 2 numerics indicating the dimensions to plot out of the selected reduction. Defaults to c(1, 2) if not specified.}

\item{viridis_color_map}{Character. A capital letter from A to H or the scale name as in \link[viridis]{scale_fill_viridis}.}

\item{verbose}{Whether to show warnings.}

\item{individual.titles}{Titles for each feature if needed. Either NULL or a vector of equal length of features.}
}
\value{
A ggplot2 object containing a Feature Plot.
}
\description{
Wrapper for \link[Seurat]{FeaturePlot}.
}
\examples{
\dontrun{
TBD
}
}
