% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_FeaturePlot.R
\name{do_FeaturePlot}
\alias{do_FeaturePlot}
\title{Wrapper for \link[Seurat]{FeaturePlot}.}
\usage{
do_FeaturePlot(
  sample,
  features,
  assay = NULL,
  reduction = NULL,
  slot = NULL,
  split.by = NULL,
  split.by.idents = NULL,
  cells.highlight = NULL,
  idents.highlight = NULL,
  dims = c(1, 2),
  pt.size = 0.5,
  fontsize = 14,
  legend = TRUE,
  legend.position = "bottom",
  legend.framewidth = 1.5,
  legend.tickwidth = 1.5,
  legend.length = 20,
  legend.width = 1,
  legend.framecolor = "grey50",
  legend.tickcolor = "white",
  plot.title = NULL,
  plot.subtitle = NULL,
  plot.caption = NULL,
  individual.titles = NULL,
  individual.subtitles = NULL,
  individual.captions = NULL,
  ncol = NULL,
  viridis_color_map = "D",
  raster = FALSE,
  raster.dpi = 1024,
  verbose = TRUE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Features to plot. It can be a single one or a vector of multiple features. Similar behavior as with \link[Seurat]{FeaturePlot}.}

\item{assay}{Assay to use. Defaults to the current assay.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}. Defaults to "umap" if present or to the last computed reduction if the argument is not provided.}

\item{slot}{Data slot to use. Character. Only one of: counts, data, scale.data. Defaults to "data".}

\item{split.by}{Split the plot in as many unique values stored in the provided metadata column.}

\item{split.by.idents}{Vector of identities to plot. The gradient scale will also be subset to only the values of such identities.}

\item{cells.highlight}{Vector of cells for which the FeaturePlot should focus into. The rest of the cells will be grayed out.}

\item{idents.highlight}{Vector of identities that the FeaturePlot should focus into. Has to match the current Seurat identities in \code{Seurat::Idents(sample)}.}

\item{dims}{Vector of 2 numerics indicating the dimensions to plot out of the selected reduction. Defaults to c(1, 2) if not specified.}

\item{pt.size}{Point size.}

\item{fontsize}{Base fontsize of the figure.}

\item{legend}{Whether to plot the legend or not.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.framewidth, legend.tickwidth}{Width of the lines of the box in the legend.}

\item{legend.length, legend.width}{Length and width of the legend. Will adjust automatically depending on legend side.}

\item{legend.framecolor, legend.tickcolor}{Color of the lines of the box in the legend.}

\item{plot.title, plot.subtitle, plot.caption}{Title for the plot or subtitle or caption.}

\item{individual.titles, individual.subtitles, individual.captions}{Titles or subtitles. for each feature if needed. Either NULL or a vector of equal length of features.}

\item{ncol}{Number of columns to use in the arrangement of the output if more than one feature is queried to the function.}

\item{viridis_color_map}{Character. A capital letter from A to H or the scale name as in \link[viridis]{scale_fill_viridis}.}

\item{raster}{Whether to raster the resulting plot. This is recommendable if plotting a lot of cells.}

\item{raster.dpi}{Numeric. Pixel resolution for rasterized plots. Defaults to 512, as per default \code{Seurat::FeaturePlot()} behavior.}

\item{verbose}{Whether to show warnings.}
}
\value{
A ggplot2 object containing a Feature Plot.
}
\description{
Wrapper for \link[Seurat]{FeaturePlot}.
}
\examples{
\dontrun{
  # Regular FeaturePlot.
  p <- SCpubr::do_FeaturePlot(sample = sample,
                              features = "nCount_RNA")

  # Add a title.
  p <- SCpubr::do_FeaturePlot(sample = sample,
                              features = "nCount_RNA",
                              plot.title = "Number of UMIs")

  # Add a subtitle.
  p <- SCpubr::do_FeaturePlot(sample = sample,
                              features = "nCount_RNA",
                              plot.title = "Number of UMIs",
                              plot.subtitle = "Number of unique mRNAs per cell.")

  # Add a caption
  p <- SCpubr::do_FeaturePlot(sample = sample,
                              features = "nCount_RNA",
                              plot.title = "Number of UMIs",
                              plot.subtitle = "Number of unique mRNAs per cell.",
                              plot.caption = "Plot generated using SCpubr.")

  # Plot multiple features and control the output columns.
  p <- SCpubr::do_FeaturePlot(sample, features = c("nCount_RNA",
                                                   "nFeature_RNA",
                                                   "percent.mt",
                                                   "CD14"),
                              plot.title = "My very important features",
                              ncol = 2)

  # Plot multiple features and control each panel plot title.
  p <- SCpubr::do_FeaturePlot(sample = sample,
                              features = c("nCount_RNA",
                                           "nFeature_RNA",
                                           "percent.mt",
                                           "CD14"),
                              plot.title = "My very important features",
                              individual.titles = c("Plot A",
                                                    "Plot_B",
                                                    NA,
                                                    "Plot_D"),
                              ncol = 2)

  # FeaturePlot with a subset of cells maintaining the original UMAP shape.
  cells.plot <- colnames(sample[, !(sample$seurat_clusters \%in\% c("2", "5", "8"))])
  p <- SCpubr::do_FeaturePlot(sample = sample,
                              cells.highlight = cells.plot,
                              features = c("CD14"))

  # FeaturePlot with a subset of identities
  # (in Seurat::Idents(sample)) maintaining the original UMAP shape.
  idents.use <- levels(sample)[!(levels(sample) \%in\% c("2", "5", "8"))]
  p <- SCpubr::do_FeaturePlot(sample = sample,
                              idents.highlight = idents.use,
                              features = c("CD14"))

  # Splitting the FeaturePlot by a variable and
  # maintaining the color scale and the UMAP shape.
  p <- SCpubr::do_FeaturePlot(sample = sample,
                              features = "CD14",
                              split.by = "seurat_clusters")

  # Splitting the FeaturePlot by a variable
  # and subset only some of the unique groups.
  p <- SCpubr::do_FeaturePlot(sample,
                              features = c("CD14", "nCount_RNA", "PC_1"),
                              split.by = "seurat_clusters",
                              split.by.idents = c("2", "5"))

  # Modify the viridis color maps.
  p <- SCpubr::do_FeaturePlot(sample = sample,
                              features = "CD14",
                              viridis_color_map = "A",
                              plot.title = "Magma")



}
}
