% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_LigandReceptorPlot.R
\name{do_LigandReceptorPlot}
\alias{do_LigandReceptorPlot}
\title{Visualize Ligand-Receptor analysis output.}
\usage{
do_LigandReceptorPlot(
  liana_output = NULL,
  split.by = NULL,
  keep_source = NULL,
  keep_target = NULL,
  top_interactions = 25,
  significance_threshold = 0.05,
  dot_border = TRUE,
  border.color = "black",
  x_labels_angle = 0,
  rotate_strip_text = FALSE,
  legend.position = "bottom",
  legend.type = "colorbar",
  legend.length = 20,
  legend.width = 1,
  legend.framecolor = "grey50",
  legend.tickcolor = "white",
  legend.framewidth = 1.5,
  legend.tickwidth = 1.5,
  viridis_color_map = "G",
  font.size = 14,
  dot.size = 1,
  font.type = "sans",
  flip = FALSE,
  plot_grid = FALSE,
  grid.color = "grey90",
  grid.type = "dotted"
)
}
\arguments{
\item{liana_output}{Object resulting from running \link[liana]{liana_wrap}.}

\item{split.by}{Character. Whether to further facet the plot on the y axis by common ligand.complex or receptor.complex. Values to provide: NULL, ligand.complex, receptor.complex.}

\item{keep_source, keep_target}{Character. Identities to keep for the source/target of the interactions. NULL otherwise.}

\item{top_interactions}{Numeric. Number of unique interactions to retrieve ordered by magnitude and specificity. It does not necessarily mean that the output will contain as many, but rather an approximate value.}

\item{significance_threshold}{Numeric. Value to filter the interactions by significance. Default is 0.05.}

\item{dot_border}{Logical. Whether to draw a black border in the dots.}

\item{border.color}{Character. Color for the border of the dots.}

\item{x_labels_angle}{Numeric. One of 0 (horizontal), 45 (diagonal), 90 (vertical). Adjusts to 0 if flip = FALSE and 45 if flip = TRUE.}

\item{rotate_strip_text}{Logical. Whether the text in the strips should be flipped 90 degrees.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.type}{Character. Type of legend to display. One of: normal, colorbar, colorsteps.}

\item{legend.length, legend.width}{Length and width of the legend. Will adjust automatically depending on legend side.}

\item{legend.framecolor, legend.tickcolor}{Color of the lines of the box in the legend.}

\item{legend.framewidth, legend.tickwidth}{Width of the lines of the box in the legend.}

\item{viridis_color_map}{Character. A capital letter from A to H or the scale name as in \link[viridis]{scale_fill_viridis}.}

\item{font.size}{Overall font.size of the plot.}

\item{dot.size}{Numeric. Size aesthetic for the dots.}

\item{font.type}{Character. Base font for the plot. One of mono, serif or sans.}

\item{flip}{Logical. Whether to invert the axis.}

\item{plot_grid}{Logical. Whether to plot grid lines.}

\item{grid.color}{Character. Color of the grid in the panels.}

\item{grid.type}{Character. One of the possible linetype options: blank, solid, dashed, dotted, dotdash, longdash, twodash.}
}
\value{
A ggplot2 plot with the results of the Ligand-Receptor analysis.
}
\description{
This function makes use of \href{https://github.com/saezlab/liana}{liana} package to run Ligand-Receptor analysis. Takes the output of liana and generates a dot-plot visualization according to the user's specifications.
}
\examples{
\dontrun{
TBD
}
}
