% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_Nebulosa_plot.R
\name{do_Nebulosa_plot}
\alias{do_Nebulosa_plot}
\title{Wrapper for Nebulosa::plot_density in Seurat.}
\usage{
do_Nebulosa_plot(
  sample,
  features,
  joint = FALSE,
  plot.title = NULL,
  legend.position = "right",
  reduction = "umap"
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Features to plot density for. It can be a single one or a vector of multiple features.}

\item{joint}{Whether to plot different features as joint density.}

\item{plot.title}{Title to use in the plot.}

\item{legend.position}{Position of the legend in the plot.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}.}
}
\value{
A ggplot2 object containing a Nebulosa plot.
}
\description{
Wrapper for Nebulosa::plot_density in Seurat.
}
\examples{
\dontrun{
TBD
}
}
