% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_NebulosaPlot.R
\name{do_NebulosaPlot}
\alias{do_NebulosaPlot}
\title{Wrapper for Nebulosa::plot_density in Seurat.}
\usage{
do_NebulosaPlot(
  sample,
  features,
  slot = NULL,
  dims = c(1, 2),
  size = 1,
  reduction = NULL,
  combine = TRUE,
  method = c("ks", "wkde"),
  joint = FALSE,
  plot.title = NULL,
  individual.titles = NULL,
  shape = 16,
  legend = TRUE,
  fontsize = 14,
  legend.position = "right",
  viridis_color_map = "D",
  verbose = TRUE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Features to plot density for. It can be a single one or a vector of multiple features.}

\item{slot}{Slot to retrieve the data from. Defaults to "data".}

\item{dims}{Vector of 2 dims to plot the data. By default, first and second from the specified reduction.}

\item{size}{Size of the dots.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}.}

\item{combine}{Whether to create a single plot out of multiple features.}

\item{method}{Kernel density estimation method. Either "ks" or "wkde" or both. See \link[Nebulosa]{plot_density} for more details.}

\item{joint}{Whether to plot different features as joint density.}

\item{plot.title}{Title to use in the plot.}

\item{individual.titles}{Titles for each feature if needed. Either NULL or a vector of equal length of features.}

\item{shape}{Shape of the geometry (ggplot number).}

\item{legend}{Whether to plot the legend or not. Logical.}

\item{fontsize}{Base fontsize of the plot.}

\item{legend.position}{Position of the legend in the plot.}

\item{viridis_color_map}{Character. A capital letter from A to H or the scale name as in \link[viridis]{scale_fill_viridis}.}

\item{verbose}{Whether to show warnings.}
}
\value{
A ggplot2 object containing a Nebulosa plot.
}
\description{
Wrapper for Nebulosa::plot_density in Seurat.
}
\examples{
\dontrun{
TBD
}
}
