% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_NebulosaPlot.R
\name{do_NebulosaPlot}
\alias{do_NebulosaPlot}
\title{Wrapper for Nebulosa::plot_density in Seurat.}
\usage{
do_NebulosaPlot(
  sample,
  features,
  slot = NULL,
  dims = c(1, 2),
  pt.size = 1,
  reduction = NULL,
  combine = TRUE,
  method = c("ks", "wkde"),
  joint = FALSE,
  return_only_joint = NULL,
  plot.title = NULL,
  plot.subtitle = NULL,
  plot.caption = NULL,
  individual.titles = NULL,
  individual.subtitles = NULL,
  individual.captions = NULL,
  shape = 16,
  legend = TRUE,
  legend.type = "colorbar",
  legend.framewidth = 1.5,
  legend.tickwidth = 1.5,
  legend.length = 20,
  legend.width = 1,
  legend.framecolor = "grey50",
  legend.tickcolor = "white",
  font.size = 14,
  font.type = "sans",
  legend.position = "bottom",
  plot_cell_borders = TRUE,
  border.size = 1.5,
  border.color = "black",
  viridis_color_map = "D",
  verbose = TRUE
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Features to plot density for. It can be a single one or a vector of multiple features.}

\item{slot}{Slot to retrieve the data from. Defaults to "data".}

\item{dims}{Vector of 2 dims to plot the data. By default, first and second from the specified reduction.}

\item{pt.size}{Size of the dots.}

\item{reduction}{Reduction to use. Can be the canonical ones such as "umap", "pca", or any custom ones, such as "diffusion". If you are unsure about which reductions you have, use \code{Seurat::Reductions(sample)}.}

\item{combine}{Whether to create a single plot out of multiple features.}

\item{method}{Kernel density estimation method. Either "ks" or "wkde" or both. See \link[Nebulosa]{plot_density} for more details.}

\item{joint}{Whether to plot different features as joint density.}

\item{return_only_joint}{Whether to only return the joint density panel. Logical.}

\item{plot.title, plot.subtitle, plot.caption}{Title to use in the plot.}

\item{individual.titles, individual.subtitles, individual.captions}{Titles, subtitles and captions for each feature if needed. Either NULL or a vector of equal length of features.}

\item{shape}{Shape of the geometry (ggplot number).}

\item{legend}{Whether to plot the legend or not. Logical.}

\item{legend.type}{Character. Type of legend to display. One of: normal, colorbar, colorsteps.}

\item{legend.framewidth, legend.tickwidth}{Width of the lines of the box in the legend.}

\item{legend.length, legend.width}{Length and width of the legend. Will adjust automatically depending on legend side.}

\item{legend.framecolor, legend.tickcolor}{Color of the lines of the box in the legend.}

\item{font.size}{Base font.size of the plot.}

\item{font.type}{Character. Base font for the plot. One of mono, serif or sans.}

\item{legend.position}{Position of the legend in the plot.}

\item{plot_cell_borders}{Logical. Whether to plot border around cells.}

\item{border.size}{Numeric. Width of the border of the cells.}

\item{border.color}{Character. Color to use for the border of the cells.}

\item{viridis_color_map}{Character. A capital letter from A to H or the scale name as in \link[viridis]{scale_fill_viridis}.}

\item{verbose}{Whether to show warnings.}
}
\value{
A ggplot2 object containing a Nebulosa plot.
}
\description{
Wrapper for Nebulosa::plot_density in Seurat.
}
\examples{
\dontrun{
  # Basic Nebulosa plot.
  p <- SCpubr::do_NebulosaPlot(sample = sample,
                              features = "CD14")

  # Querying multiple features.
  p <- SCpubr::do_NebulosaPlot(sample,
                               features = c("CD14", "CD8A"))

  # Compute joint density.
  p <- SCpubr::do_NebulosaPlot(sample = sample,
                               features = c("CD14", "CD8A"),
                               joint = TRUE)

  # Return only the joint density panel and modify the plot.title.
  p <- SCpubr::do_NebulosaPlot(sample = sample,
                               features = c("CD14", "CD8A"),
                               joint = TRUE)

  # Modify individual titles and also add a general one.
  # Use NA when you don't want to modify the panel title.
  p <- SCpubr::do_NebulosaPlot(sample = sample,
                               features = c("CD14", "CD8A"),
                               joint = TRUE,
                               individual.titles = c("Plot A",
                                                     NA,
                                                     "Combined density"),
                               plot.title = "Density analysis")

  # Modify viridis color maps.
  p <- SCpubr::do_NebulosaPlot(sample = sample,
                               features = "CD8A",
                               viridis_color_map = "A",
                               plot.title = "Magma")

}
}
