% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_PseudotimePlot.R
\name{do_PseudotimePlot}
\alias{do_PseudotimePlot}
\title{Compute Pseudotime analysis plots.}
\usage{
do_PseudotimePlot(
  sample,
  cds,
  compute_monocle_partitions = TRUE,
  compute_monocle_clusters = FALSE,
  trajectory_graph_color = "black",
  trajectory_graph_segment_size = 1,
  pseudotime_genes = NULL,
  is_max_score_the_start = TRUE,
  label_roots = FALSE,
  label_branches = FALSE,
  label_leaves = FALSE,
  pt.size = 1,
  border.size = 1.5,
  rotate_x_axis_labels = F,
  legend.position = "bottom",
  legend.type = "colorbar",
  font.size = 14,
  font.type = "sans",
  legend.length = 30,
  legend.width = 1,
  legend.framewidth = 1.5,
  legend.tickwidth = 1.5,
  legend.framecolor = "grey50",
  legend.tickcolor = "white",
  viridis_color_map = "D"
)
}
\arguments{
\item{sample}{Seurat object.}

\item{cds}{Cell Data Set of the same seurat object. Can be obtained using \code{SeuratWrappers::as.cell_data_set(sample)}.}

\item{compute_monocle_partitions}{Whether to tell monocle3 to compute different partitions. FALSE will treat all the UMAP as a single partition.}

\item{compute_monocle_clusters}{Whether to make monocle3 to re-compute clustering}

\item{trajectory_graph_color}{Character. Color of the trajectory graph plotted on top of the UMAP.}

\item{trajectory_graph_segment_size}{Integer. Size of the trajectory graph.}

\item{pseudotime_genes}{Character. List of genes that will be used to compute enrichment scores, that will be used for pseudotime.}

\item{is_max_score_the_start}{Logical. Do the cells with the highest enrichment scores depict the beginning of the trajectory (TRUE) or the end (FALSE)?}

\item{label_roots, label_branches, label_leaves}{Logical. Label roots, branches or leaves in the trajectory graph.}

\item{pt.size}{Numeric. Point size for the plots.}

\item{border.size}{Numeric. Point size for the border of the cells in the plots.}

\item{rotate_x_axis_labels}{Logical. Whether to rotate X axis labels in the}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{legend.type}{Character. Type of legend to display. One of: normal, colorbar, colorsteps.}

\item{font.size}{Numeric. Overall fontsize for the plots.}

\item{font.type}{Character. Base font for the plot. One of mono, serif or sans.}

\item{legend.length, legend.width}{Length and width of the legend. Will adjust automatically depending on legend side.}

\item{legend.framewidth, legend.tickwidth}{Width of the lines of the box in the legend.}

\item{legend.framecolor, legend.tickcolor}{Color of the lines of the box in the legend.}

\item{viridis_color_map}{Character. Viridis color map to use in the FeaturePlot and dotplot for the enrichment scores.}
}
\value{
A list containing a collection of ggplot2 objects.
}
\description{
Compute Pseudotime analysis plots.
}
\examples{
\dontrun{
TBD
}
}
