% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_TermEnrichmentPlot.R
\name{do_TermEnrichmentPlot}
\alias{do_TermEnrichmentPlot}
\title{Compute the enriched terms for a given list of genes.}
\usage{
do_TermEnrichmentPlot(
  genes,
  dbs_use,
  nchar_wrap = 20,
  nterms = 10,
  font.size = 14,
  font.type = "sans",
  site = "Enrichr",
  plot.title = NULL,
  plot.subtitle = NULL,
  plot.caption = NULL,
  legend.position = "bottom",
  legend.type = "colorbar",
  colors.use = NULL,
  text_labels_size = 4,
  legend.length = 30,
  legend.width = 1,
  legend.framewidth = 1.5,
  legend.tickwidth = 1.5,
  legend.framecolor = "grey50",
  legend.tickcolor = "white"
)
}
\arguments{
\item{genes}{Character vector with the genes to query.}

\item{dbs_use}{Character vector with the desired EnrichR databases to query. This can be consulted with \code{enrichR::listEnrichrDbs()}. Alternatively, one of the following values:
\itemize{
\item "A": Will perform a query to 4 databases for cell types (Azimuth, Descartes, PanglaoDB and Descartes) and 4 databases for functional terms (MsigDB, GO-BP, GO-MF and KEGG). This is the default option if this parameter is not provided.
\item "B": Performs a query for the cell type databases (Azimuth, Descartes, PanglaoDB and Descartes).
\item "C": Performs a query for the functional terms (MsigDB, GO-BP, GO-MF and KEGG).
}}

\item{nchar_wrap}{Number of characters to use as a limit to wrap the term names. The higher this value, the longer the lines would be for each term in the plots. Defaults to 60.}

\item{nterms}{Number of terms to report for each database. Terms are arranged by adjusted p-value and selected from lowest to highest. Defaults to 5.}

\item{font.size}{Base font size for the plot. Defaults to 14.}

\item{font.type}{Character. Base font for the plot. One of mono, serif or sans.}

\item{site}{Site to query the genes against. Can be one of: "Enrichr", "FlyEnrichr", "WormEnrichr", "YeastEnrichr", "FishEnrichr".}

\item{plot.title, plot.subtitle, plot.caption}{Title,subtitle or caption to use in the plot.}

\item{legend.position}{Position of the legend in the plot.}

\item{legend.type}{Character. Type of legend to display. One of: normal, colorbar, colorsteps. Colorsteps will default to colorbar if only one value is present for the scale.}

\item{colors.use}{Character vector of 2 colors (low and high ends of the color scale) to generate the gradient.}

\item{text_labels_size}{Numeric. Controls how big or small labels are in the plot.}

\item{legend.length, legend.width}{Length and width of the legend. Will adjust automatically depending on legend side.}

\item{legend.framewidth, legend.tickwidth}{Width of the lines of the box in the legend.}

\item{legend.framecolor, legend.tickcolor}{Color of the lines of the box in the legend.}
}
\value{
A ggplot2 object with enriched terms.
}
\description{
Compute the enriched terms for a given list of genes.
}
\examples{
\dontrun{
  genes <- c("ABCB1", "ABCG2", "AHR", "AKT1", "AR")
  # General use case.
  p <- SCpubr::do_TermEnrichmentPlot(genes = genes,
                                     dbs_use = "C",
                                     ncol = 2)
  p

  # Modify total number of terms displayed.
  p <- SCpubr::do_TermEnrichmentPlot(genes = genes,
                                     dbs_use = "GO_Biological_Process_2021",
                                     nterms = 15)
  p

  # Modify the length of the terms displayed.
  p <- SCpubr::do_TermEnrichmentPlot(genes = genes,
                                     dbs_use = "GO_Biological_Process_2021",
                                     nterms = 15,
                                     nchar_wrap = 30)
  p
}
}
