% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_VlnPlot.R
\name{do_VlnPlot}
\alias{do_VlnPlot}
\title{Wrapper for \link[Seurat]{VlnPlot}.}
\usage{
do_VlnPlot(
  sample,
  features,
  assay = NULL,
  slot = NULL,
  group.by = NULL,
  split.by = NULL,
  legend = FALSE,
  colors.use = NULL,
  pt.size = 0,
  y_cut = NULL,
  plot_boxplot = TRUE,
  legend.position = "bottom",
  plot.title = NULL,
  individual.titles = NULL,
  xlab = NULL,
  ylab = NULL,
  fontsize = 14,
  ncol = NULL,
  legend.ncol = 3,
  rotate_x_labels = NULL
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Features to represent.}

\item{assay}{Assay to use. Defaults to the current assay.}

\item{slot}{Data slot to use. Character. Only one of: counts, data, scale.data. Defaults to "data".}

\item{group.by}{Variable you want the cells to be colored for.}

\item{split.by}{CURRENTLY NOT WORKING. Split into as many plots as unique values in the variable provided.}

\item{legend}{Whether to plot the legend or not.}

\item{colors.use}{Named vector of colors matching the unique values in either the current identities in the Seurat object of the unique values in group.by or split.by.}

\item{pt.size}{Size of points in the VlnPlot.}

\item{y_cut}{Vector with the values in which the Violins should be cut. Only works for one feature.}

\item{plot_boxplot}{Logical. Whether to plot a Box plot inside the violin or not.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{plot.title}{Title to use in the plot.}

\item{individual.titles}{Titles for each feature if needed. Either NULL or a vector of equal length of features.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{fontsize}{Base fontsize for the figure.}

\item{ncol}{Numeric. Number of columns to arrange multiple plots into.}

\item{legend.ncol}{Number of columns in the legend.}

\item{rotate_x_labels}{Logical. Whether to rotate X axis labels to horizontal or not. If multiple features, a vector of logical values of the same length.}
}
\value{
A ggplot2 object containing a Violin Plot.
}
\description{
Wrapper for \link[Seurat]{VlnPlot}.
}
\examples{
\dontrun{
TBD
}
}
