% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_VlnPlot.R
\name{do_VlnPlot}
\alias{do_VlnPlot}
\title{Wrapper for \link[Seurat]{VlnPlot}.}
\usage{
do_VlnPlot(
  sample,
  features,
  assay = NULL,
  slot = NULL,
  group.by = NULL,
  split.by = NULL,
  legend = FALSE,
  colors.use = NULL,
  pt.size = 0,
  y_cut = NULL,
  plot_boxplot = TRUE,
  legend.position = "bottom",
  plot.title = NULL,
  individual.titles = NULL,
  xlab = NULL,
  ylab = NULL,
  axis.text.fonsize = 16,
  axis.title.fonsize = 16,
  plot.title.fonsize = 18,
  legend.text.fontsize = 16,
  legend.title.fontsize = 16,
  ncol = NULL,
  legend.ncol = 3
)
}
\arguments{
\item{sample}{Seurat object.}

\item{features}{Features to represent.}

\item{assay}{Assay to use. Defauls to the current assay.}

\item{slot}{Data slot to use. Character. Only one of: counts, data, scale.data. Defaults to "data".}

\item{group.by}{Variable you want the cells to be colored for.}

\item{split.by}{CURRENTLY NOT WORKING. Split into as many plots as unique values in the variable provided.}

\item{legend}{Whether to plot the legend or not.}

\item{colors.use}{Named vector of colors matching the unique values in either the current identities in the seurat object of the unique values in group.by or split.by.}

\item{pt.size}{Size of points in the VlnPlot.}

\item{y_cut}{Vector with the values in which the Violins should be cut. Only works for one feature.}

\item{plot_boxplot}{Logical. Whether to plot a boxplot inside the violin or not.}

\item{legend.position}{Position of the legend in the plot. Will only work if legend is set to TRUE.}

\item{plot.title}{Title to use in the plot.}

\item{individual.titles}{Titles for each feature if needed. Either NULL or a vector of equal length of features.}

\item{xlab}{Title for the X axis.}

\item{ylab}{Title for the Y axis.}

\item{axis.text.fonsize}{Modify the fontsize for axis texts.}

\item{axis.title.fonsize}{Modify the fontsize for axis titles.}

\item{plot.title.fonsize}{Modify the fontsize for the plot title.}

\item{legend.text.fontsize}{Modify the fontsize for the legend text.remove_x_axis}

\item{legend.title.fontsize}{Modify the fontsize for the legend title.}

\item{ncol}{Numeric. Number of columns to arrange multiple plots into.}

\item{legend.ncol}{Number of columns in the legend.}
}
\value{
A ggplot2 object containing a Violin Plot.
}
\description{
Wrapper for \link[Seurat]{VlnPlot}.
}
\examples{
\dontrun{
TBD
}
}
