% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_report}
\alias{package_report}
\title{Generate a status report of SCpubr and its dependencies.}
\usage{
package_report(startup = FALSE)
}
\arguments{
\item{startup}{\strong{\code{\link[base]{logical}}} | Whether the message should be displayed at startup, therefore, also containing welcoming messages and tips. If \strong{\code{FALSE}}, only the report itself will be printed.}
}
\value{
None
}
\description{
This function generates a summary report of the installation status of SCpubr, which packages are still missing and which functions can or can not currently be used.
}
\examples{

\donttest{
# Print a package report.
SCpubr::package_report()
}
}
