% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_Plot.R
\name{save_Plot}
\alias{save_Plot}
\title{Save a plot as png, pdf and svg.}
\usage{
save_Plot(
  plot,
  figure_path = NULL,
  create_path = TRUE,
  file_name = NULL,
  dpi = 300,
  output_format = "publication",
  width = 8,
  height = 8
)
}
\arguments{
\item{plot}{Plot to save.}

\item{figure_path}{\strong{\code{\link[base]{character}}} | Path where the figure will be stored.}

\item{create_path}{\strong{\code{\link[base]{logical}}} | Whether to create the path.}

\item{file_name}{\strong{\code{\link[base]{character}}} | Name of the file (without extension, it will be added automatically).}

\item{dpi}{\strong{\code{\link[base]{numeric}}} | Dpi to use.}

\item{output_format}{\strong{\code{\link[base]{character}}} | Output format of the saved figure.  One of:
\itemize{
\item \emph{\code{pdf}}: Saves the figure as a PDF file.
\item \emph{\code{png}}: Saves the figure as a PNG file.
\item \emph{\code{jpeg}}: Saves the figure as a JPEG file.
\item \emph{\code{tiff}}: Saves the figure as a TIFF file.
\item \emph{\code{svg}}: Saves the figure as a SVG file.
\item \emph{\code{publication}}: Saves the figure as PDF, PNG and SVG files.
\item \emph{\code{all}}: Saves the figure in all possible formats.
}}

\item{width, height}{\strong{\code{\link[base]{numeric}}} | Width and height of the figure (inches).}
}
\value{
Nothing.
}
\description{
Save a plot as png, pdf and svg.
}
\examples{
\dontrun{
  # Generate a plot.
  p <- SCpubr::do_DimPlot(sample = sample)

  # Default parameters.
  SCpubr::savePlot(plot = p)

  # Specifying the name and folder.
  SCpubr::savePlot(plot = p,
                   figure_path = "/path/to/my/figures/",
                   file_name = "my_figure")

  # Specify to also create a new folder.
  SCpubr::savePlot(plot = p,
                   figure_path = "/path/to/my/figures/",
                   file_name = "my_figure",
                   create_path = TRUE)

  # Set dimensions for the figure.
  SCpubr::savePlot(plot = p,
                   figure_path = "/path/to/my/figures/",
                   file_name = "my_figure",
                   create_path = TRUE,
                   width = 8,
                   height = 8)

  # Set quality (dpi).
  SCpubr::savePlot(plot = p,
                   figure_path = "/path/to/my/figures/",
                   file_name = "my_figure",
                   create_path = TRUE,
                   width = 8,
                   height = 8,
                   dpi = 300)
}
}
