% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{state_dependencies}
\alias{state_dependencies}
\title{State SCpubr current function dependencies.}
\usage{
state_dependencies(func_name = NULL)
}
\arguments{
\item{func_name}{Name of an exported function from SCpubr. If NULL, return all functions.}
}
\value{
None
}
\description{
State SCpubr current function dependencies.
}
\examples{
\dontrun{
TBD
}
}
