% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{state_dependencies}
\alias{state_dependencies}
\title{State SCpubr current function dependencies.}
\usage{
state_dependencies(function_name = NULL, return_dependencies = FALSE)
}
\arguments{
\item{function_name}{\strong{\code{\link[base]{character}}} | Name of an exported function from SCpubr. If NULL, return all functions.}

\item{return_dependencies}{\strong{\code{\link[base]{logical}}} | Whether to have the dependencies as an output object instead of a printed message.}
}
\value{
None
}
\description{
State SCpubr current function dependencies.
}
\examples{

# See all dependencies.
SCpubr::state_dependencies()

# See the dependencies for a single package.
SCpubr::state_dependencies(function_name = "do_DimPlot")
}
